% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{adjustment_codes}
\alias{adjustment_codes}
\title{Adjustment Codes}
\usage{
adjustment_codes(
  df = NULL,
  col = NULL,
  type = c("all", "carc", "rarc"),
  action = c("review", "join"),
  ...
)
}
\arguments{
\item{df}{data.frame}

\item{col}{column of Adjustment codes to match on}

\item{type}{type of Adjustment code; \code{all} (default), \code{carc}, \code{rarc}}

\item{action}{action to take; \code{review} (default), \code{join}}

\item{...}{Empty}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
CARC and RARC Codes
}
\section{Claim Adjustment Reason Codes}{


\emph{X12 External Code Source 139}

These codes describe why a claim or service line was paid differently
than it was billed and generally assign responsibility for the
adjustment amounts.

The Claim Adjustment Group Codes (e.g., PR, OA) are internal to the X12
standard. The format is always two alpha characters:
\itemize{
\item CO: Contractual Obligations
\item CR: Corrections and Reversals
\item OA: Other Adjustments
\item PI: Payer Initiated Reductions
\item PR: Patient Responsibility
}
}

\section{Remittance Advice Remark Codes}{


\emph{X12 External Code Source 411}

These codes provide additional explanation for an adjustment already
described by a Claim Adjustment Reason Code (CARC) or convey information
about remittance processing.

Remittance Advice Remark Codes (RARCs) are used to provide additional
explanation for an adjustment already described by a Claim Adjustment Reason
Code (CARC) or to convey information about remittance processing.

There are two types of RARCs, supplemental and informational. The majority
of the RARCs are supplemental; these are generally referred to as RARCs
without further distinction.

Supplemental RARCs provide additional explanation for an adjustment already
described by a CARC.

The second type of RARC is informational; these RARCs are all prefaced
with \verb{Alert:} and are often referred to as Alerts.

Alerts are used to convey information about remittance processing and are
never related to a specific adjustment or CARC.
}

\examples{
adjustment_codes()$group

adjustment_codes(type = "carc",
                 action = "review") |>
                 head()

adjustment_codes(type = "rarc",
                 action = "review") |>
                 head()

dplyr::tibble(code = c("CO-253", "OA-23", "PI-185")) |>
adjustment_codes(col = "code",
                 type = "carc",
                 action = "join")
}
