% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{assign_adjustments}
\alias{assign_adjustments}
\title{Assign Adjustment Codes}
\usage{
assign_adjustments(code, include_keys = FALSE, ...)
}
\arguments{
\item{code}{\verb{<chr>} vector of adjustment codes; should be of the form
\code{GROUP-CARC}, where \code{GROUP} is two letters, followed by a dash (\code{-}) and
\code{CARC} is a two-to-three character alphanumeric string.}

\item{include_keys}{\verb{<lgl>} include keys in output; default is \code{FALSE}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Assign Adjustment Codes
}
\examples{
x <- c("CO-253", "OA-23", "PI-185", "-45")

x

assign_adjustments(x)

assign_adjustments(x,
  include_keys = TRUE)

dplyr::tibble(code = x,
  desc = assign_adjustments(code))

dplyr::tibble(code = x) |>
  dplyr::mutate(desc = purrr::map(
  code,
  assign_adjustments))

purrr::map_df(x,
  assign_adjustments)

purrr::map_df(x,
  assign_adjustments,
  include_keys = TRUE)

dplyr::tibble(code = x,
  purrr::map_dfr(code, assign_adjustments))

assign_adjustments(x, include_keys = TRUE) |>
  as.data.frame() |>
  dplyr::tibble()

}
