% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_amounts}
\alias{calc_amounts}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
calc_amounts(wrvu, prvu_f, prvu_nf, mrvu, wgpci, pgpci, mgpci, cf)
}
\arguments{
\item{wrvu}{numeric}

\item{prvu_f}{numeric}

\item{prvu_nf}{numeric}

\item{mrvu}{numeric}

\item{wgpci}{numeric}

\item{pgpci}{numeric}

\item{mgpci}{numeric}

\item{cf}{numeric}
}
\value{
description
}
\description{
Calculate Physician Fee Schedule Payment Amounts
}
\examples{
calc_amounts(wrvu     = 6.26,
             prvu_nf  = 7.92,
             prvu_f   = 4.36,
             mrvu     = 0.99,
             wgpci    = 1.0,
             pgpci    = 0.883,
             mgpci    = 1.125,
             cf       = 32.7442)
}
