% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_limiting_charge}
\alias{calc_limiting_charge}
\title{Calculate Limiting Charge}
\usage{
calc_limiting_charge(participating_amount)
}
\arguments{
\item{participating_amount}{numeric}
}
\value{
description
}
\description{
Calculate Limiting Charge
}
\examples{
calc_limiting_charge(26.35)
}
