% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_nonpar_amount}
\alias{calc_nonpar_amount}
\title{Calculate Non-Participating Amount}
\usage{
calc_nonpar_amount(participating_amount)
}
\arguments{
\item{participating_amount}{numeric}
}
\value{
description
}
\description{
Calculate Non-Participating Amount
}
\examples{
calc_nonpar_amount(26.35)
}
