% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calculate_amounts}
\alias{calculate_amounts}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
calculate_amounts(wrvu, fprvu, nprvu, mrvu, wgpci, pgpci, mgpci, cf = 32.7442)
}
\arguments{
\item{wrvu}{\emph{< dbl >} Work RVU}

\item{fprvu}{\emph{< dbl >} Facility Practice Expense RVU}

\item{nprvu}{\emph{< dbl >} Non-Facility Practice Expense RVU}

\item{mrvu}{\emph{< dbl >} Malpractice RVU}

\item{wgpci}{\emph{< dbl >} Work GPCI}

\item{pgpci}{\emph{< dbl >} Practice Expense GPCI}

\item{mgpci}{\emph{< dbl >} Malpractice GPCI}

\item{cf}{\emph{< dbl >} Conversion Factor, default is \code{32.7442}}
}
\value{
A list (invisibly) of the Participating, Non-Participating &
Limiting Charge Amounts for both Facility & Non-Facility RVUs
}
\description{
((wRVU x wGPCI) + (pRVU x pGPCI) + (mRVU x mGPCI)) x Conversion Factor
}
\examples{
calculate_amounts(wrvu  = 6.26,
                  nprvu = 7.92,
                  fprvu = 4.36,
                  mrvu  = 0.99,
                  wgpci = 1.053,
                  pgpci = 0.883,
                  mgpci = 1.125,
                  cf    = 32.7442)
}
