% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{carc_add_dash}
\alias{carc_add_dash}
\title{Format CARC Codes for Identification}
\usage{
carc_add_dash(x, placeholder = "||")
}
\arguments{
\item{x}{\verb{<chr>} vector of CARC adjustment codes; should be of the form
\code{GROUP-CARC}, where \code{GROUP} is two letters, followed by a dash (\code{-}) and
\code{CARC} is a two-to-three character alphanumeric string.}

\item{placeholder}{\verb{<chr>} placeholder string for missing elements of CARC
codes; default is \code{||}}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Format CARC Codes for Identification
}
\examples{
carc_add_dash(c("- 253", "OA-23", "PI-", "-45 ", "OA23"))

}
\keyword{internal}
