% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_asc}
\alias{case_asc}
\title{Replace HCPCS Level II ASC Group Indicator Descriptions}
\usage{
case_asc(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of ASC Group indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{asc_description} column
}
\description{
Replace HCPCS Level II ASC Group Indicator Descriptions
}
\examples{
dplyr::tibble(asc_grp = c("YY", "99")) |> case_asc(asc_grp)
}
