% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_assistant_surgery}
\alias{case_assistant_surgery}
\title{Add Assistant Surgery Descriptions}
\usage{
case_assistant_surgery(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Assistant Surgery indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{asst_description} column
}
\description{
Fee schedule amount equals 16 percent of amount
otherwise applicable for surgical payment.
}
\details{
Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}
}
\examples{
dplyr::tibble(surg_asst = c(0:2, "9")) |> case_assistant_surgery(surg_asst)
}
\keyword{internal}
