% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_bilateral}
\alias{case_bilateral}
\title{Add Bilateral Surgery Descriptions}
\usage{
case_bilateral(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Bilateral Surgery indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{bilat_description} column
}
\description{
Modifier 50: Indicates services subject to payment adjustment
}
\examples{
dplyr::tibble(surg_bilat = c(0:3, "9")) |> case_bilateral(surg_bilat)
}
\keyword{internal}
