% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling.R
\name{case_chapter}
\alias{case_chapter}
\title{Add HCPCS Level I & II Section Labels}
\usage{
case_chapter(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{chapter} column
}
\description{
Add HCPCS Level I & II Section Labels
}
\examples{
dplyr::tibble(
hcpcs = c("39503", "99215", "99140",
          "69990", "70010", "0222U",
          "V5299", "7010F", "0074T")) |>
          case_chapter(hcpcs)

}
