% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{case_chapter_icd10}
\alias{case_chapter_icd10}
\title{Add ICD-10-CM Chapter Labels}
\usage{
case_chapter_icd10(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of ICD-10-CM codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{chapter} column
}
\description{
Add ICD-10-CM Chapter Labels
}
\examples{
dplyr::tibble(
code = c("F50.8", "G40.311", "Q96.8", "Z62.890", "R45.4", "E06.3")) |>
case_chapter_icd10(code)
}
