% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_co_surgeon}
\alias{case_co_surgeon}
\title{Add Co-Surgeons Descriptions}
\usage{
case_co_surgeon(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Co Surgeon indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{cosurg_description} column
}
\description{
Modifier 62: Services for which two surgeons,
each in a different specialty, may be paid.
}
\examples{
dplyr::tibble(surg_co = c(0:2, "9")) |> case_team_surgery(surg_co)
}
\keyword{internal}
