% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_cpt_category}
\alias{case_cpt_category}
\title{Add CPT Category Labels}
\usage{
case_cpt_category(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{cpt_category} column
}
\description{
Add CPT Category Labels
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
x <- c("39503", "99215", "99140", "69990", "70010",
       "0222U", "V5299", "7010F", "0074T")
dplyr::tibble(hcpcs = x) |>
case_cpt_category(hcpcs)
\dontshow{\}) # examplesIf}
}
\keyword{internal}
