% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_global}
\alias{case_global}
\title{Add Global Days Descriptions}
\usage{
case_global(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Global Days indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{global_description} column
}
\description{
Add Global Days Descriptions
}
\examples{
dplyr::tibble(global = c("000", "010", "090",
                         "MMM", "XXX", "YYY", "ZZZ")) |>
case_global(global)
}
