% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_global_days}
\alias{case_global_days}
\title{Add Global Days Descriptions}
\usage{
case_global_days(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Global Days indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{global_description} column
}
\description{
Add Global Days Descriptions
}
\examples{
dplyr::tibble(global_days = c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ")) |>
case_global_days(global_days)
}
\keyword{internal}
