% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_hcpcs_level}
\alias{case_hcpcs_level}
\title{Add HCPCS Level Labels}
\usage{
case_hcpcs_level(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{hcpcs_level} column
}
\description{
Add HCPCS Level Labels
}
\examples{
x <- c("39503", "99215", "99140", "69990", "70010",
       "0222U", "V5299", "7010F", "0074T")
dplyr::tibble(hcpcs = x) |>
case_hcpcs_level(hcpcs)
}
\keyword{internal}
