% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_imaging}
\alias{case_imaging}
\title{Add Diagnostic Imaging Descriptions}
\usage{
case_imaging(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Diagnostic Imaging indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{dximg_description} column
}
\description{
Identifies the applicable Diagnostic Service family for
HCPCS codes with a Multiple Procedure indicator of 4.
}
\examples{
dplyr::tibble(dximg = stringr::str_pad(c(1:11, 88, 99),
width = "2", pad = "0")) |>
case_imaging(dximg)
}
