% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level2.R
\name{case_labcert}
\alias{case_labcert}
\title{Add HCPCS Level II Lab Certification Descriptions}
\usage{
case_labcert(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Lab Certification indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{labcert_description} column
}
\description{
Add HCPCS Level II Lab Certification Descriptions
}
\examples{
x <- c("010", 100, 110, 115, 120, "110, 120, 130, 400")
dplyr::tibble(labcert = x) |> case_labcert(labcert)
}
