% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling.R
\name{case_level}
\alias{case_level}
\title{Add HCPCS Level Labels}
\usage{
case_level(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{level} column
}
\description{
Add HCPCS Level Labels
}
\examples{
dplyr::tibble(
      hcpcs = c("39503", "99215", "99140",
                "69990", "70010", "0222U",
                "V5299", "7010F", "0074T")) |>
                case_level(hcpcs)
}
