% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_modifier}
\alias{case_modifier}
\title{Add Modifier Indicator Descriptions}
\usage{
case_modifier(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Modifier indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with an \code{mod_description} column
}
\description{
Add Modifier Indicator Descriptions
}
\examples{
dplyr::tibble(mod = c(26, "TC", 53)) |> case_modifier(mod)
}
