% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_multiple_pricing}
\alias{case_multiple_pricing}
\title{Add HCPCS Level II Multiple Pricing Indicator Descriptions}
\usage{
case_multiple_pricing(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Multiple Pricing indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{mult_description} column
}
\description{
Add HCPCS Level II Multiple Pricing Indicator Descriptions
}
\examples{
dplyr::tibble(mult_pi = c(9, LETTERS[1:7])) |> case_multiple_pricing(mult_pi)
}
