% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_multproc}
\alias{case_multproc}
\title{Add Multiple Procedure Descriptions}
\usage{
case_multproc(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Multiple Procedure indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{mproc_description} column
}
\description{
Modifier 51: Indicates applicable payment adjustment rule for multiple procedures
}
\examples{
dplyr::tibble(mult_proc = as.character(0:9)) |>
case_multproc(mult_proc)
}
