% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling.R
\name{case_opps}
\alias{case_opps}
\title{Add OPPS Indicator Descriptions}
\usage{
case_opps(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of OPPS Indicator indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with an \code{opps_description} column
}
\description{
Add OPPS Indicator Descriptions
}
\examples{
dplyr::tibble(opps_ind = c("1", "9")) |>
case_opps(opps_ind)

}
