% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_opps_ind}
\alias{case_opps_ind}
\title{Add OPPS Indicator Descriptions}
\usage{
case_opps_ind(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of OPPS Indicator indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with an \code{opps_description} column
}
\description{
Add OPPS Indicator Descriptions
}
\examples{
dplyr::tibble(opps_ind = c("1", "9")) |> case_opps_ind(opps_ind)
}
\keyword{internal}
