% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_pctc}
\alias{case_pctc}
\title{Add PCTC Indicator Descriptions}
\usage{
case_pctc(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of PCTC indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{pctc_description} column
}
\description{
Add PCTC Indicator Descriptions
}
\examples{
dplyr::tibble(pctc = as.character(0:9)) |> case_pctc(pctc)
}
\keyword{internal}
