% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_physician_supervision}
\alias{case_physician_supervision}
\title{Add Physician Supervision Descriptions}
\usage{
case_physician_supervision(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Physician Supervision indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{supvis_description} column
}
\description{
This field is for use in post payment review.
}
\examples{
x <- stringr::str_pad(c(1:6, 9, 21, 22, 66, "6A", 77, "7A"),
                      width = "2", pad = "0")
dplyr::tibble(supvis = x) |>
case_physician_supervision(supvis)
}
\keyword{internal}
