% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling.R
\name{case_pricing}
\alias{case_pricing}
\title{Add HCPCS Level II Pricing Indicator Descriptions}
\usage{
case_pricing(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Pricing indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{price_description} column
}
\description{
Add HCPCS Level II Pricing Indicator Descriptions
}
\examples{
dplyr::tibble(price = c("00", 11:13, 21:22, 31:46, 51:57, 99)) |>
case_pricing(price)
}
