% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_section_hcpcs}
\alias{case_section_hcpcs}
\title{Add HCPCS Section Labels}
\usage{
case_section_hcpcs(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{section} column
}
\description{
Add HCPCS Section Labels
}
\examples{
dplyr::tibble(
      hcpcs = c("39503", "99215", "99140",
                "69990", "70010", "0222U",
                "V5299", "7010F", "0074T")) |>
                case_section_hcpcs(hcpcs)
}
