% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{case_section_icd10}
\alias{case_section_icd10}
\title{Add ICD-10-CM Section Labels}
\usage{
case_section_icd10(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of HCPCS codes to match on}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{section} column
}
\description{
Add ICD-10-CM Section Labels
}
\examples{
dplyr::tibble(code = c("F50.8", "G40.311", "Q96.8",
                       "Z62.890", "R45.4", "E06.3")) |>
                       case_section_icd10(code)
}
