% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_status}
\alias{case_status}
\title{Add Status Code Descriptions}
\usage{
case_status(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Status Codes}
}
\value{
A \link[tibble:tibble-package]{tibble} with an \code{status_description} column
}
\description{
Add Status Code Descriptions
}
\examples{
dplyr::tibble(status = LETTERS) |> case_status(status)
}
\keyword{internal}
