% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_team}
\alias{case_team}
\title{Add Team Surgery Descriptions}
\usage{
case_team(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Team Surgery indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{team_description} column
}
\description{
Modifier 66: Services for which Team Surgeons may be paid
}
\examples{
dplyr::tibble(surg_team = c(0:2, "9")) |> case_team(surg_team)
}
\keyword{internal}
