% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/relabeling.R
\name{case_tos}
\alias{case_tos}
\title{Add HCPCS Level II Type of Service Descriptions}
\usage{
case_tos(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Type of Service codes}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{type_of_service} column
}
\description{
Add HCPCS Level II Type of Service Descriptions
}
\examples{
dplyr::tibble(tos = c(0:9, LETTERS)) |> case_tos(tos)
}
