% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{cases}
\alias{cases}
\alias{case_section}
\alias{case_section_cpt}
\alias{case_section_hcpcs}
\alias{case_level}
\alias{case_category}
\alias{case_global}
\alias{case_imaging}
\alias{case_supervision}
\alias{case_team}
\alias{case_bilateral}
\alias{case_multproc}
\alias{case_cosurg}
\alias{case_assistant}
\alias{case_opps}
\alias{case_modifier}
\alias{case_pctc}
\alias{case_status}
\alias{case_asc}
\alias{case_coverage}
\alias{case_pricing}
\alias{case_multiple_pricing}
\alias{case_labcert}
\alias{case_tos}
\title{Recoding Functions}
\usage{
case_section(df, col)

case_section_cpt(df, col)

case_section_hcpcs(df, col)

case_level(df, col)

case_category(df, col)

case_global(df, col)

case_imaging(df, col)

case_supervision(df, col)

case_team(df, col)

case_bilateral(df, col)

case_multproc(df, col)

case_cosurg(df, col)

case_assistant(df, col)

case_opps(df, col)

case_modifier(df, col)

case_pctc(df, col)

case_status(df, col, desc = FALSE)

case_asc(df, col)

case_coverage(df, col)

case_pricing(df, col)

case_multiple_pricing(df, col)

case_labcert(df, col)

case_tos(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Type of Service codes}

\item{desc}{add Descriptions}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{section} column

A \link[tibble:tibble-package]{tibble} with a \code{section} column

A \link[tibble:tibble-package]{tibble} with a \code{section} column

A \link[tibble:tibble-package]{tibble} with a \code{level} column

A \link[tibble:tibble-package]{tibble} with a \code{category} column

A \link[tibble:tibble-package]{tibble} with a \code{global_description} column

A \link[tibble:tibble-package]{tibble} with a \code{dximg_description} column

A \link[tibble:tibble-package]{tibble} with a \code{supvis_description} column

A \link[tibble:tibble-package]{tibble} with a \code{team_description} column

A \link[tibble:tibble-package]{tibble} with a \code{bilat_description} column

A \link[tibble:tibble-package]{tibble} with a \code{mproc_description} column

A \link[tibble:tibble-package]{tibble} with a \code{cosurg_description} column

A \link[tibble:tibble-package]{tibble} with a \code{asst_description} column

A \link[tibble:tibble-package]{tibble} with an \code{opps_description} column

A \link[tibble:tibble-package]{tibble} with an \code{mod_description} column

A \link[tibble:tibble-package]{tibble} with a \code{pctc_description} column

A \link[tibble:tibble-package]{tibble} with a \code{status_desc} column

A \link[tibble:tibble-package]{tibble} with a \code{asc_description} column

A \link[tibble:tibble-package]{tibble} with a \code{cov_description} column

A \link[tibble:tibble-package]{tibble} with a \code{price_description} column

A \link[tibble:tibble-package]{tibble} with a \code{mult_description} column

A \link[tibble:tibble-package]{tibble} with a \code{labcert_description} column

A \link[tibble:tibble-package]{tibble} with a \code{tos_description} column
}
\description{
Various recoding functions

Identifies the applicable Diagnostic Service family for
HCPCS codes with a Multiple Procedure indicator of 4.

This field is for use in post payment review.

Modifier 66: Services for which Team Surgeons may be paid

Modifier 50: Indicates services subject to payment adjustment

Modifier 51: Indicates applicable payment adjustment rule for multiple procedures

Modifier 62: Services for which two surgeons,
each in a different specialty, may be paid.

Fee schedule amount equals 16 percent of amount
otherwise applicable for surgical payment.
}
\details{
Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
cd <- dplyr::tibble(
      hcpcs = c("39503", "99215", "99140",
                "69990", "70010", "0222U",
                "V5299", "7010F", "0074T")

cd |> case_section(hcpcs)
cd |> case_section_cpt(hcpcs)
cd |> case_section_hcpcs(hcpcs)
cd |> case_level(hcpcs)
cd |> case_category(hcpcs)

# Global Days
glob <- c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ")
dplyr::tibble(global = glob) |>
case_global(global)


# Diagnostic Imaging Family
di <- stringr::str_pad(
      c(1:11, 88, 99), width = "2", pad = "0")
dplyr::tibble(dximg = di) |>
case_imaging(dximg)

# Physician Supervision
sv <- stringr::str_pad(
      c(1:6, 9, 21:22, 66, "6A", 77, "7A"), width = "2", pad = "0")
dplyr::tibble(supvis = sv) |>
case_supervision(supvis)

# Team Surgery
dplyr::tibble(surg_team = c(0:2, "9")) |>
case_team(surg_team)

# Bilateral Surgery
dplyr::tibble(surg_bilat = c(0:3, "9")) |>
case_bilateral(surg_bilat)

# Multiple Procedure
dplyr::tibble(mult_proc = as.character(0:9)) |>
case_multproc(mult_proc)

# Co-Surgeons
dplyr::tibble(surg_co = c(0:2, "9")) |>
case_cosurg(surg_co)

# Assistant at Surgery
dplyr::tibble(surg_asst = c(0:2, "9")) |>
case_assistant(surg_asst)

# OPPS Indicator
dplyr::tibble(opps_ind = c("1", "9")) |>
case_opps(opps_ind)

# Modifier Indicator
dplyr::tibble(mod = c(26, "TC", 53)) |>
case_modifier(mod)

# PCTC Indicator
dplyr::tibble(pctc = as.character(0:9)) |>
case_pctc(pctc)

# Status Codes
dplyr::tibble(status = LETTERS) |>
case_status(status, desc = TRUE)
\dontshow{\}) # examplesIf}
dplyr::tibble(asc_grp = c("YY", "99")) |> case_asc(asc_grp)
dplyr::tibble(cov = c("C", "D", "I", "M", "S")) |> case_coverage(cov)
dplyr::tibble(price = c("00", 11:13, 21:22, 31:46, 51:57, 99)) |>
case_pricing(price)
dplyr::tibble(mult_pi = c(9, LETTERS[1:7])) |> case_multiple_pricing(mult_pi)
x <- c("010", 100, 110, 115, 120, "110, 120, 130, 400")
dplyr::tibble(labcert = x) |> case_labcert(labcert)
dplyr::tibble(tos = c(0:9, LETTERS)) |> case_tos(tos)
}
