% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{chr}
\alias{chr}
\title{Alias for \code{as.character()}}
\usage{
chr(...)
}
\arguments{
\item{...}{arguments to pass to \code{as.character()}}
}
\description{
Alias for \code{as.character()}
}
\keyword{internal}
