% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{compare_addons}
\alias{compare_addons}
\title{Return a HCPCS Code's Add-On Code Complement}
\usage{
compare_addons(hcpcs, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{Empty dots}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Return a HCPCS Code's Add-On Code Complement
}
\examples{
compare_addons(c("22633", "0074T"))

compare_addons(c("22633", "22630", "22532",
                 "77001", "88334", "64727"))

compare_addons(c("11646", "0074T", "39503",
                 "43116", "33935", "11646"))

}
