% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regex.R
\name{construct_regex}
\alias{construct_regex}
\title{Construct regex patterns}
\usage{
construct_regex(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<chr>} vector
}
\description{
Construct regex patterns
}
\examples{
construct_regex(search_descriptions()$hcpcs_code)

construct_regex(search_hcpcs()$hcpcs_code)

# Incorrect:
construct_regex(search_adjustments()$adj_code)

# Should be:
"^[A-DM-PWY1-9][AIOR0-9]?[0-9]{0,3}?$"

# Test adj codes
c("4", "CO", "P6", "100", "B19", "MA22", "MA124", "N766")

}
