% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{data_dict}
\alias{data_dict}
\title{Data dictionary for PFS and RVU data}
\usage{
data_dict(out = c("md", "df"), dict = c("pfs", "rvu"))
}
\arguments{
\item{out}{description}

\item{dict}{description}
}
\value{
description
}
\description{
Data dictionary for PFS and RVU data
}
\examples{
data_dict()

}
\keyword{internal}
