% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{describe_hcpcs}
\alias{describe_hcpcs}
\title{Describe HCPCS Codes}
\usage{
describe_hcpcs(hcpcs = NULL, desc_type = "All", varname = "hcpcs", ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{desc_type}{\verb{<chr>} vector of code description types; \code{All} (default),
\code{Short}, \code{Long}, \code{Medium}, \code{Medical}, \code{Consumer}, \code{Clinician}, \verb{Proprietary Name}}

\item{varname}{If \code{hcpcs} is a \link{data.frame} or a
\link[tibble:tibble-package]{tibble}, this is the quoted name of the column
containing HCPCS codes; default is \code{"hcpcs"}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Describe HCPCS Codes
}
\examples{
describe_hcpcs(hcpcs = c("39503", "43116", "33935", "11646"))

describe_hcpcs(
  dplyr::tibble(
    hcpcs = c("A0021", "V5362", "J9264", "G8916")))

}
