% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{descriptors}
\alias{descriptors}
\title{2023 CPT Descriptors (Clinician & Consumer-Friendly)}
\usage{
descriptors(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2023 CPT Descriptors (Clinician & Consumer-Friendly)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
descriptors(hcpcs = c("39503", "43116", "33935", "11646"))
\dontshow{\}) # examplesIf}
}
