% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addons.R
\name{get_aoc_type}
\alias{get_aoc_type}
\title{HCPCS Add-On code type}
\usage{
get_aoc_type(hcpcs, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{Empty dots}
}
\value{
\verb{<list>} of four \verb{<chr>} vectors indicating group membership:
\code{primary}, \code{addon}, \code{both}, and \code{neither}
}
\description{
HCPCS Add-On code type
}
\examples{
get_aoc_type(c("11646", "0074T"))

c("22633", "22630", "22532", "77001", "88334", "0715T", "64727") |>
get_aoc_type()

}
