% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_descriptions}
\alias{get_descriptions}
\title{Get HCPCS Descriptions}
\usage{
get_descriptions(hcpcs = NULL, desc_type = "All", ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{desc_type}{\verb{<chr>} vector of description types; \code{All} (default),
\code{Short}, \code{Long}, \code{Medium}, \code{Medical}, \code{Consumer}, \code{Clinician},
\verb{Proprietary Name}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get HCPCS Descriptions
}
\examples{
get_descriptions(hcpcs = c("39503", "43116", "33935", "11646"))

get_descriptions(hcpcs = c("A0021", "V5362", "J9264", "G8916"))

}
