% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_example}
\alias{get_example}
\title{Load Example Datasets}
\usage{
get_example(name = c("report", "practicum"))
}
\arguments{
\item{name}{name of example}
}
\value{
\verb{<tibble>}
}
\description{
Load Example Datasets
}
\keyword{internal}
