% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phys_sched.R
\name{get_gpcis}
\alias{get_gpcis}
\title{Physician Fee Schedule Geographic Practice Cost Indices}
\usage{
get_gpcis(mac = NULL, state = NULL, locality = NULL, ...)
}
\arguments{
\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{state}{\verb{<chr>} vector of state abbreviations}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Physician Fee Schedule Geographic Practice Cost Indices
}
\examples{
get_gpcis(state = "GA", locality = "01", mac = "10212")

}
