% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{get_hcpcs_type}
\alias{get_hcpcs_type}
\title{HCPCS code type}
\usage{
get_hcpcs_type(hcpcs_code, ...)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \verb{<list>} of three \verb{<chr>} vectors indicating set membership:
\code{cpt}, \code{hcpcs}, and \code{neither}
}
\description{
HCPCS code type
}
\examples{
# Only Category I CPTs
get_hcpcs_type(hcpcs = c("39503", "43116", "33935", "11646"))

# Only Level II HCPCS
get_hcpcs_type(hcpcs = c("A0021", "V5362", "J9264", "G8916"))

# Both
get_hcpcs_type(hcpcs = c("39503", "43116", "A0021", "V5362"))

# Neither
get_hcpcs_type(hcpcs = c("CCCCC", "0002U", "0003U", "0004U", "1164F", "0074T"))

# All Three
get_hcpcs_type(hcpcs = c("39503", "43116", "A0021", "V5362", "1164F", "0074T"))

}
