% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{get_pin}
\alias{get_pin}
\title{Get a pinned dataset from a \link[pins:pins-package]{pins} board}
\usage{
get_pin(pin, ...)
}
\arguments{
\item{pin}{\verb{<chr>} string name of pinned dataset}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
\verb{<tibble>}
}
\description{
Get a pinned dataset from a \link[pins:pins-package]{pins} board
}
