% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncci.R
\name{get_ptp_edits}
\alias{get_ptp_edits}
\title{Get NCCI Procedure-to-Procedure (PTP) Edits}
\usage{
get_ptp_edits(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Get NCCI Procedure-to-Procedure (PTP) Edits
}
\examples{
get_ptp_edits(hcpcs = c("43116"))

}
