% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rbcs.R
\name{get_rbcs_type}
\alias{get_rbcs_type}
\title{RBCS-HCPCS code type}
\usage{
get_rbcs_type(hcpcs, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
A \verb{<list>} of three \verb{<chr>} vectors indicating set membership:
\code{rbcs_cpt}, \code{rbcs_hcpcs}, \code{rbcs_only} and \code{rbcs_not}
}
\description{
RBCS-HCPCS code type
}
\examples{
# RBCS and Category I CPTs
get_rbcs_type(hcpcs = c("39503", "43116", "33935", "11646"))

# RBCS and Level II HCPCS
get_rbcs_type(hcpcs = c("J9264"))

# RBCS and Both
get_rbcs_type(hcpcs = c("39503", "J9264"))

# RBCS Only
get_rbcs_type(hcpcs = c("0002U", "0003U", "0004U"))

# None
get_rbcs_type(hcpcs = c("CCCCC", "1164F", "0074T"))

# All Four
get_rbcs_type(hcpcs = c("39503", "J9264", "0002U", "A0021"))

}
