% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{github_raw}
\alias{github_raw}
\title{Return GitHub raw url}
\usage{
github_raw(x)
}
\arguments{
\item{x}{url}
}
\value{
raw url
}
\description{
Return GitHub raw url
}
\examples{
github_raw("andrewallenbruce/provider/")
}
