% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gt.R
\name{gt_style}
\alias{gt_style}
\title{gt Theme}
\usage{
gt_style(gt_tbl, lbl = TRUE, tablign = "center", tabsize = 16)
}
\arguments{
\item{gt_tbl}{description}

\item{lbl}{description}

\item{tablign}{description}

\item{tabsize}{description}
}
\value{
description
}
\description{
gt Theme
}
\keyword{internal}
