% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-gt.R
\name{gt_theme_northstar}
\alias{gt_theme_northstar}
\title{{gt} Theme}
\usage{
gt_theme_northstar(
  gt_object,
  column_labels = TRUE,
  tab_align = "center",
  tab_size = 16,
  tab_weight = "bold",
  ...
)
}
\arguments{
\item{gt_object}{\verb{<gt_tbl>} A \link[gt:gt-package]{gt} table object}

\item{column_labels}{\verb{<lgl>} Show column labels, default is \code{TRUE}}

\item{tab_align}{\verb{<chr>} Stub text alignment, default is \code{center}}

\item{tab_size}{\verb{<int>} Font size, default is \code{16}}

\item{tab_weight}{Font weight, default is \code{bold}}

\item{...}{Additional arguments to \link[gt:gt-package]{gt::table_options()}}
}
\value{
An object of class \verb{<gt_tbl>}
}
\description{
{gt} Theme
}
\keyword{internal}
