% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{hcpcs}
\alias{hcpcs}
\title{2024 Healthcare Common Procedure Coding System (HCPCS)}
\usage{
hcpcs()
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2024 Healthcare Common Procedure Coding System (HCPCS)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hcpcs()
\dontshow{\}) # examplesIf}
}
