% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{hcpcs_lv2}
\alias{hcpcs_lv2}
\title{2024 Healthcare Common Procedure Coding System (HCPCS)}
\usage{
hcpcs_lv2(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2024 Healthcare Common Procedure Coding System (HCPCS)
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hcpcs_lv2(hcpcs = c("39503", "43116", "33935", "11646"))
\dontshow{\}) # examplesIf}
}
