% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{hcpcs_search}
\alias{hcpcs_search}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
hcpcs_search(hcpcs, state = NULL, locality = NULL, mac = NULL)
}
\arguments{
\item{hcpcs}{numeric}

\item{state}{numeric}

\item{locality}{numeric}

\item{mac}{numeric}
}
\value{
description
}
\description{
Calculate Physician Fee Schedule Payment Amounts
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hcpcs_search(hcpcs = "V5299",
             state = "GA",
             locality = "99",
             mac = "10212")

hcpcs_search(hcpcs = c("39503", "43116", "33935", "11646"),
             state = "GA")
\dontshow{\}) # examplesIf}
}
