% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10api}
\alias{icd10api}
\title{Search the NLM's ICD-10-CM API}
\source{
National Institute of Health/National Library of Medicine
}
\usage{
icd10api(code = NULL, term = NULL, field = c("both", "code"), limit = 500, ...)
}
\arguments{
\item{code}{All or part of an ICD-10-CM code}

\item{term}{Associated term describing an ICD-10 code}

\item{field}{options are "code" or "both"; default is "both"}

\item{limit}{API limit is 500; defaults to 10}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} containing the search results.
}
\description{
\code{\link[=icd10api]{icd10api()}} allows you to search the National Library of
Medicine's ICD-10-CM API by code or associated term.
}
\details{
ICD-10-CM (International Classification of Diseases, 10th Revision,
Clinical Modification) is a medical coding system for classifying
diagnoses and reasons for visits in U.S. health care settings.
\subsection{Links}{
\itemize{
\item \href{https://clinicaltables.nlm.nih.gov/apidoc/icd10cm/v3/doc.html}{NIH NLM Clinical Table Service ICD-10-CM API}
\item \href{http://www.cdc.gov/nchs/icd/icd10cm.htm}{Learn more about ICD-10-CM.}
}
}
}
\note{
Current Version: ICD-10-CM \strong{2024}
}
