% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd_sections}
\alias{icd_sections}
\title{ICD-10-CM Section Labels and Regexes}
\usage{
icd_sections(code = NULL)
}
\arguments{
\item{code}{vector of ICD-10-CM codes}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
ICD-10-CM Section Labels and Regexes
}
\examples{
icd_sections(c("F50.8", "G40.311", "Q96.8", "Z62.890", "R45.4",
               "E06.3", "H00.019", "D50.1", "C4A.70", "Z20.818"))
}
