% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{is_carc_group}
\alias{is_carc_group}
\title{Validate CARC Group Codes}
\usage{
is_carc_group(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of CARC adjustment codes; should be of the form
\code{GROUP-CARC}, where \code{GROUP} is two letters, followed by a dash (\code{-}) and
\code{CARC} is a two-to-three character alphanumeric string.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Validate CARC Group Codes
}
\examples{
x <- c("- 253", "OA-23", "PI-", "-45 ", "OA23")

is_carc_group(x)

x[which(is_carc_group(x))]

}
