% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_category_I}
\alias{is_category_I}
\title{Check if code is HCPCS Level 1 Category I (CPT)}
\usage{
is_category_I(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
Check if code is HCPCS Level 1 Category I (CPT)
}
\examples{
is_category_I("11646")

is_category_I("1164F")
}
