% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_category_II}
\alias{is_category_II}
\title{Check if code is HCPCS Level 1 Category II}
\usage{
is_category_II(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
Category II: Supplemental tracking codes that can be used for performance
measurement. These codes are intended to facilitate data collection about
quality of care by coding certain services and/or test results that support
performance measures and that have been agreed upon as contributing to
good patient care.
}
\details{
Some codes in this category may relate to compliance by the health care
professional with state or federal law. The use of these codes is optional.
The codes are not required for correct coding and may not be used as a
substitute for Category I codes. Services/procedures or test results
described in this category make use of alpha characters as the 5th
character in the string (i.e., 4 digits followed by an alpha character).

These digits are not intended to reflect the placement of the code in the
regular (Category I) part of the CPT code set. Also, these codes describe
components that are typically included in an evaluation and management
service or test results that are part of the laboratory test/procedure.

Consequently, they do not have a relative value associated with them.

Category II codes are:
\itemize{
\item Alphanumeric and consist of four digits followed by the letter F
\item NOT billing codes
\item Used to track services on claims for performance measurement
\item Not to be used as a substitute for Category I codes
}
}
\examples{
is_category_II("1164F")

try(is_category_II("11646"))
}
