% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_cpt_category_I}
\alias{is_cpt_category_I}
\title{Validate HCPCS Level I (CPT) Category I Code}
\usage{
is_cpt_category_I(hcpcs_code)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
A valid CPT Category I code is:
}
\details{
\itemize{
\item 5 characters long
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with one of \verb{[AMU0-9]}
}
}
\examples{
x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_cpt_category_I(x)

x[which(is_cpt_category_I(x))]

}
