% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_cpt_category_III}
\alias{is_cpt_category_III}
\title{Validate HCPCS Level I Category III Code}
\source{
\href{https://www.ama-assn.org/practice-management/cpt/category-iii-codes}{AMA Category III Codes}
}
\usage{
is_cpt_category_III(hcpcs_code)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
A valid CPT Category III code is:
}
\details{
\itemize{
\item 5 characters long
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with a \verb{[T]}
}

CPT Category III codes are a set of temporary codes that allow data
collection for emerging technologies and are used in the the Food and Drug
Administration (FDA) approval process.

No RVUs are assigned to these codes and payment is based on payer policy. If
a Category III code is available, it must be reported in place of a Category
I unlisted code.

The procedure/service they describe may not meet the following Category I
requirements:
\itemize{
\item Necessary devices/drugs have received FDA clearance/approval.
\item It is performed by many US physicians/QHPs, at a frequency consistent with intended clinical use.
\item It is consistent with current medical practice.
\item It's clinical efficacy is documented in CPT-approved literature.
}
}
\examples{
x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_cpt_category_I(x)

x[which(is_cpt_category_III(x))]

}
