% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_hcpcs_level_I}
\alias{is_hcpcs_level_I}
\title{Validate HCPCS Level I (CPT) Code}
\usage{
is_hcpcs_level_I(hcpcs_code)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
A valid HCPCS Level I code is:
}
\details{
\itemize{
\item 5 characters long
\item Begins with any 4 digits, \verb{[0-9]\{4\}}, and
\item Ends with one of \verb{[AFMTU0-9]}
}

HCPCS Level I is comprised of CPT (Current Procedural Terminology),
a numeric coding system maintained by the American Medical Association
(AMA).

CPT is a uniform coding system consisting of descriptive terms and
identifying codes that are used primarily to identify medical services and
procedures furnished by physicians and other health care professionals.

These health care professionals use the CPT to identify services and
procedures for which they bill public or private health insurance programs.

Decisions regarding the addition, deletion, or revision of CPT codes are
made by the AMA. The CPT codes are republished and updated annually by the
AMA.

Level I does not include codes needed to separately report medical items or
services that are regularly billed by suppliers other than physicians.
}
\examples{
x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_hcpcs_level_I(x)

x[which(is_hcpcs_level_I(x))]

}
