% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_hcpcs_level_II}
\alias{is_hcpcs_level_II}
\title{Validate HCPCS Level II Code}
\usage{
is_hcpcs_level_II(hcpcs_code)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
A valid HCPCS Level II code is:
}
\details{
\itemize{
\item 5 characters long
\item Begins with one of \verb{[A-CEGHJ-MP-V]} and
\item Ends with any 4 digits, \verb{[0-9]\{4\}}
}

Level II of the HCPCS is a standardized coding system that is used primarily
to identify products, supplies, and services not included in the Level I CPT
codes.

These include items such as ambulance services and durable medical equipment,
prosthetics, orthotics, and supplies (DMEPOS) when used outside a physician's
office.

Because Medicare and other insurers cover a variety of services, supplies,
and equipment that are not identified by CPT codes, the level II HCPCS codes
were established for submitting claims for these items.

Level II codes are also referred to as alpha-numeric codes because they
consist of a single alphabetical letter followed by 4 numeric digits, while
CPT codes are identified using 5 numeric digits.
}
\examples{
x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_hcpcs_level_II(x)

x[which(is_hcpcs_level_II(x))]

}
