% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_level_I}
\alias{is_level_I}
\title{Check if code is HCPCS Level I}
\usage{
is_level_I(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
HCPCS Level I is comprised of CPT (Current Procedural Terminology),
a numeric coding system maintained by the American Medical Association
(AMA). CPT is a uniform coding system consisting of descriptive terms and
identifying codes that are used primarily to identify medical services and
procedures furnished by physicians and other health care professionals.
These health care professionals use the CPT to identify services and
procedures for which they bill public or private health insurance programs.
}
\details{
Decisions regarding the addition, deletion, or revision of CPT codes are
made by the AMA. The CPT codes are republished and updated annually by the
AMA. Level I of the HCPCS, the CPT codes, does not include codes needed to
separately report medical items or services that are regularly billed by
suppliers other than physicians.
}
\examples{
is_level_I("11646")

is_level_I("E8015")

purrr::map_vec(c("11646", "E8015"), is_level_I)

}
