% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_level_II}
\alias{is_level_II}
\title{Check if code is HCPCS Level II}
\usage{
is_level_II(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{\verb{<chr>} string of \code{length(1)}}

\item{arg}{\verb{<chr>} function argument name in the current function}

\item{call}{\verb{<environment>} environment the function is called from}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, \code{FALSE} otherwise
}
\description{
HCPCS Level II is a standardized coding system that is used primarily to
identify products, supplies, and services not included in the CPT code set
jurisdiction, such as ambulance services and durable medical equipment,
prosthetics, orthotics, and supplies (DMEPOS) when used outside a
physician's office.
}
\details{
Level II of the HCPCS is a standardized coding system that is used primarily
to identify products, supplies, and services not included in the CPT codes,
such as ambulance services and durable medical equipment, prosthetics,
orthotics, and supplies (DMEPOS) when used outside a physician's office.
Because Medicare and other insurers cover a variety of services, supplies,
and equipment that are not identified by CPT codes, the level II HCPCS codes
were established for submitting claims for these items. The development and
use of level II of the HCPCS began in the 1980's. Level II codes are also
referred to as alpha-numeric codes because they consist of a single
alphabetical letter followed by 4 numeric digits, while CPT codes are
identified using 5 numeric digits.
}
\examples{
is_level_II("E8015")

is_level_II("11646")

}
