% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{is_rarc_code}
\alias{is_rarc_code}
\title{Validate RARC Codes}
\usage{
is_rarc_code(x)
}
\arguments{
\item{x}{\verb{<chr>} vector of RARC adjustment codes, a two-to-three character
alphanumeric string.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Validate RARC Codes
}
\examples{
x <- c("- 253", "OA-23", "PI-", "-45 ",
       "OA23", "MA109", "N9", "N722")

is_rarc_code(x)

x[which(is_rarc_code(x))]

}
