% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_hcpcs}
\alias{is_valid_hcpcs}
\title{Validate HCPCS Code}
\usage{
is_valid_hcpcs(hcpcs_code)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} string}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, otherwise \code{FALSE}
}
\description{
A valid HCPCS code is:
}
\details{
\itemize{
\item 5 characters long
\item Begins with one of \verb{[A-CEGHJ-MP-V0-9]}
\item Followed by any 3 digits, \verb{[0-9]\{3\}}, and
\item Ends with one of \verb{[AFMTU0-9]}
}
}
\examples{
x <- c('T1503', 'G0478', '81301', '69641', '0583F', '0779T', NA)

is_valid_hcpcs(x)

x[which(is_valid_hcpcs(x))]

try(is_valid_hcpcs("1164"))

}
