% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_icd}
\alias{is_valid_icd}
\title{Check if string is valid ICD-10-CM code}
\usage{
is_valid_icd(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
Check if string is valid ICD-10-CM code
}
\examples{
purrr::map_vec(
c("H00.019", "D50.1", "C4A.70", "Z20.818", "11646", "E8015"),
is_valid_icd)

}
