% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_length}
\alias{is_valid_length}
\title{Check if string is valid length of a HCPCS code}
\usage{
is_valid_length(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
Check if string is valid length of a HCPCS code
}
\examples{
is_valid_length("11646")

try(is_valid_length("1164"))

}
