% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd_ncd.R
\name{lcd}
\alias{lcd}
\title{Local Coverage Determinations}
\usage{
lcd(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{Empty dots}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Database Last Updated 2023-04-27
}
\examples{
lcd(hcpcs = c("A4555", "E0766"))

}
