% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{limiting_charge}
\alias{limiting_charge}
\title{Calculate limiting charge}
\usage{
limiting_charge(x)
}
\arguments{
\item{x}{numeric}
}
\value{
description
}
\description{
Calculate limiting charge
}
\examples{
limiting_charge(26.35)
}
