% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{modifiers}
\alias{modifiers}
\title{Level I and II HCPCS Modifiers}
\usage{
modifiers(mod = NULL, ...)
}
\arguments{
\item{mod}{\emph{\if{html}{\out{<chr>}}} 2-digit HCPCS modifier}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Level I and II HCPCS Modifiers
}
\examples{
modifiers(mod = c("25", "59"))
}
