% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{ncd}
\alias{ncd}
\title{NCD Download Database}
\usage{
ncd()
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Last Updated 2022-12-08
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
ncd()
\dontshow{\}) # examplesIf}
}
