% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{non_participating_amount}
\alias{non_participating_amount}
\title{Calculate Non-Participating Amount}
\usage{
non_participating_amount(par_amount)
}
\arguments{
\item{par_amount}{< \emph{double} > Participating Amount}
}
\value{
numeric vector of Non-participating amount
}
\description{
The payment amount for Non-participating physicians is 95\% of the payment
amount for Participating physicians (i.e., the fee schedule amount).
}
\examples{
non_participating_amount(26.35)
}
