% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{non_participating_fee}
\alias{non_participating_fee}
\title{Calculate Non-Participating Fee}
\usage{
non_participating_fee(participating_fee)
}
\arguments{
\item{participating_fee}{\verb{<dbl>} Participating Fee}
}
\value{
\verb{<dbl>} Non-Participating Fee
}
\description{
The payment amount for Non-participating physicians is 95\% of the payment
amount for Participating physicians (i.e., the fee schedule amount):
}
\details{
\deqn{rvu_{total} = rvu_{w}(gpci_{w}) + rvu_{pe}(gpci_{pe}) + rvu_{mp}(gpci_{mp})}
}
\examples{
non_participating_fee(26.35)
}
