% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/northstar-package.R
\docType{package}
\name{northstar-package}
\alias{northstar}
\alias{northstar-package}
\title{northstar: Tidy Revenue Integrity Tools}
\description{
Revenue integrity tools.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://andrewallenbruce.github.io/northstar/}
  \item \url{https://github.com/andrewallenbruce/northstar}
  \item Report bugs at \url{https://github.com/andrewallenbruce/northstar/issues}
}

}
\author{
\strong{Maintainer}: Andrew Bruce \email{andrew.bruce@northstarrevenueintegrity.com} [copyright holder]

}
\keyword{internal}
