% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{pfs}
\alias{pfs}
\title{2024 Physician Fee Schedule Payment Amount File}
\usage{
pfs(hcpcs = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs}{5-digit HCPCS code}

\item{mac}{5-digit Medicare Administrative Contractor (MAC) code}

\item{locality}{2-digit locality ID}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2024 Physician Fee Schedule Payment Amount File
}
\examples{
pfs(hcpcs    = c("39503", "43116", "33935", "11646"),
    locality = "01",
    mac      = "10212") |>
dplyr::glimpse()
}
