% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rbcs}
\alias{rbcs}
\title{Restructured BETOS Classification for HCPCS}
\usage{
rbcs(
  hcpcs = NULL,
  category = NULL,
  subcategory = NULL,
  family = NULL,
  procedure = NULL,
  limit_cols = TRUE
)
}
\arguments{
\item{hcpcs}{< \emph{character} > HCPCS code}

\item{category}{< \emph{character} > RBCS Category:
\itemize{
\item \code{Procedure} (n = 6920)
\item \code{Test} (n = 3015)
\item \code{DME} (n = 2971)
\item \code{Treatment} (n = 1795)
\item \code{Imaging} (n = 1097)
\item \code{E&M} (n = 695)
\item \code{Anesthesia} (n = 307)
\item \code{Other} (n = 233)
}}

\item{subcategory}{< \emph{character} > RBCS Subcategory (53 unique in total)}

\item{family}{< \emph{character} > RBCS Family (178 unique in total)}

\item{procedure}{< \emph{character} > Procedure Type:
\itemize{
\item \code{Major} (n = 3676)
\item \code{Non-Procedure} (n = 10113)
\item \code{Other} (n = 3244)
}}

\item{limit_cols}{< \emph{logical} > Limit Columns}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Column} \tab \strong{Description} \cr
   \code{hcpcs} \tab HCPCS or CPT code \cr
   \code{rbcs} \tab RBCS Identifier \cr
   \code{category} \tab RBCS Category \cr
   \code{subcategory} \tab RBCS Subcategory \cr
   \code{family} \tab RBCS Family \cr
   \code{procedure} \tab RBCS Major Procedure Indicator \cr
   \code{date_hcpcs_add} \tab Date HCPCS Code was added \cr
   \code{date_hcpcs_end} \tab Date HCPCS Code was no longer effective \cr
   \code{date_rbcs_assign} \tab Earliest Date that the RBCS ID was effective \cr
}
}
\description{
\code{\link[=rbcs]{rbcs()}} allows the user to group HCPCS codes into clinically
meaningful categories based on the original \emph{Berenson-Eggers Type of Service}
(BETOS) classification.
}
\section{From BETOS to RBCS}{


The Restructured BETOS Classification System (RBCS) is a taxonomy that allows
researchers to group Medicare Part B healthcare service codes into clinically
meaningful categories and subcategories.

Based on the original Berenson-Eggers Type of Service (BETOS) classification
created in the 1980s, it includes notable updates such as Part B non-physician
services and undergoes annual updates by a technical expert panel of
researchers and clinicians.

The general framework for grouping service codes into the new RBCS taxonomy
largely follows the same structure of BETOS. Like BETOS, the RBCS groups
HCPCS codes into categories, subcategories, and families – with categories
as the most aggregate level and families as the more granular level.

All Medicare Part B service codes, including non-physician services, are
assigned to a 6-character RBCS taxonomy code.
}

\section{Links}{

\itemize{
\item \href{https://data.cms.gov/provider-summary-by-type-of-service/provider-service-classifications/restructured-betos-classification-system}{Restructured BETOS Classification System}
\item \href{https://data.cms.gov/resources/restructured-betos-classification-system-data-dictionary}{RBCS Data Dictionary}
}
}

\section{Update Frequency}{
 Annually
}

\examples{
rbcs(hcpcs = c("J9264", "39503", "43116", "33935", "11646")) |>
dplyr::glimpse()
}
