% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{search_adjustments}
\alias{search_adjustments}
\title{Search Adjustment Codes}
\usage{
search_adjustments(adj_code = NULL, adj_type = NULL, ...)
}
\arguments{
\item{adj_code}{\verb{<chr>} vector of Adjustment codes; default is \code{NULL}}

\item{adj_type}{\verb{<chr>} type of Adjustment code; \code{Group}, \code{CARC}, \code{RARC};
default is \code{NULL}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
CARC and RARC Codes
}
\details{
Claim Adjustment Reason Codes:

\emph{X12 External Code Source 139}

These codes describe why a claim or service line was paid differently than
it was billed and generally assign responsibility for the adjustment
amounts.

The Claim Adjustment \emph{Group Codes} are internal to the X12 standard. The
format is always two alpha characters:
\itemize{
\item \strong{CO}: Contractual Obligations
\item \strong{CR}: Corrections and Reversals
\item \strong{OA}: Other Adjustments
\item \strong{PI}: Payer Initiated Reductions
\item \strong{PR}: Patient Responsibility
}

Remittance Advice Remark Codes:

\emph{X12 External Code Source 411}

Remittance Advice Remark Codes (RARCs) are used to provide additional
explanation for an adjustment already described by a Claim Adjustment
Reason Code (CARC) or to convey information about remittance processing.

There are two types of RARCs: \strong{Supplemental} and \strong{Informational}:

The majority of RARCs are \emph{supplemental} and, as such, are generally
referred to as RARCs without further distinction. Supplemental RARCs
provide additional explanation for an adjustment already described by a
CARC.

The second type is \emph{informational}; these are all prefaced with \verb{Alert:}
and are referred to as Alerts. They are used to convey information about
remittance processing and are \emph{never} related to a specific adjustment or
CARC.
}
\examples{
search_adjustments(adj_type = "Group")

search_adjustments(adj_type = "CARC")

search_adjustments(adj_type = "RARC")

}
\concept{HIPAA Code Standards}
