% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfs.R
\name{search_anesthesia}
\alias{search_anesthesia}
\title{search_anesthesia}
\usage{
search_anesthesia(mac = NULL, locality = NULL, ...)
}
\arguments{
\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
search_anesthesia
}
\examples{
search_anesthesia(locality = "01", mac = "10212")

}
