% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{search_cpt}
\alias{search_cpt}
\title{HCPCS Level I (CPT) Codes}
\usage{
search_cpt(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
HCPCS Level I (CPT) Codes
}
\examples{
search_cpt(hcpcs = c("39503", "43116", "33935", "11646"))

}
