% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustments.R
\name{search_denials}
\alias{search_denials}
\title{Search Denial Types}
\usage{
search_denials(...)
}
\arguments{
\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search Denial Types
}
\examples{
search_denials()

}
