% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_schedule.R
\name{search_fee_schedule}
\alias{search_fee_schedule}
\title{Search Physician Fee Schedule}
\usage{
search_fee_schedule(hcpcs, state = NULL, locality = NULL, mac = NULL, ...)
}
\arguments{
\item{hcpcs}{< \emph{chr} > 5-character HCPCS Code}

\item{state}{< \emph{chr} > 2-character State Abbreviation}

\item{locality}{< \emph{chr} > 2-digit Locality ID}

\item{mac}{< \emph{chr} > 5-digit MAC ID code}

\item{...}{Empty}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Column} \tab \strong{Description} \cr
   \code{hcpcs} \tab 5-character HCPCS Code \cr
   \code{hcpcs_type} \tab HCPCS Type \cr
   \code{description} \tab HCPCS Description \cr
   \code{cons_desc} \tab HCPCS Level I Consumer Description \cr
   \code{clin_descs} \tab HCPCS Level I Clinician Descriptions \cr
   \code{mod} \tab Modifier \cr
   \code{status} \tab Status Code \cr
   \code{mac} \tab Medicare Administrative Contractor ID \cr
   \code{state} \tab State Abbreviation \cr
   \code{locality} \tab State Abbreviation \cr
   \code{area} \tab State Abbreviation \cr
   \code{counties} \tab State Abbreviation \cr
   \code{two_macs} \tab State Abbreviation \cr
   \code{wgpci} \tab State Abbreviation \cr
   \code{pgpci} \tab State Abbreviation \cr
}
}
\description{
Search Physician Fee Schedule
}
\examples{
search_fee_schedule(hcpcs    = c("V5299", "70170"),
                    state    = "GA",
                    locality = "99",
                    mac      = "10212")
}
