% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfs.R
\name{search_gpcis}
\alias{search_gpcis}
\title{Geographic Practice Cost Indices (GPCIs)}
\usage{
search_gpcis(mac = NULL, state = NULL, locality = NULL, ...)
}
\arguments{
\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{state}{\verb{<chr>} vector of state abbreviations}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
A geographic practice cost index (GPCI) has been established for every
Medicare payment locality for each of the three components of a procedure's
relative value unit (RVU): work, practice expense, and malpractice. The GPCI
reflects the relative costs of physician work, practice expense, and
malpractice insurance in an area compared to the national average costs for
each component.
}
\details{
The GPCIs are applied in the calculation of a Medicare payment schedule
amount by multiplying the RVU for each component of a procedure by the GPCI
for the locality in which the service is furnished.

GPCIs allow for considerably less variation in physicians' costs of practice
than under the previous reasonable charge system. The GPCIs are based on the
premise that the costs of practice resources are similar across the country,
but that the costs of operating a medical practice can vary significantly
from one area to another.
\subsection{Physician Work (Cost-of-Living)}{

The work GPCIs are designed to reflect the relative cost of physician labor
by Medicare locality and are measured from salary information of individuals
with higher education.

To address physician workforce needs, Congress established a 1.0 floor for
for physician work RVUs and incentivizes Alaska by establishing a permanent
1.5 floor (the highest work GPCI.) The remaining work GPCIs range from 1.002
to 1.1. Because the cost-of-living GPCI (work) accounts for only one quarter
of geographic differences, the range is small.
}

\subsection{Practice Expense}{

The practice expense GPCIs are designed to measure geographic variation in
the prices of inputs to medical practice. They are comprised of four
component indices: employee wages, purchased services, office rent and
equipment, supplies, and other miscellaneous expenses.

Congress established the following states as "Frontier States": Montana,
Nevada, North Dakota, South Dakota, and Wyoming. The practice expense GPCIs
for these states are set at a 1.0 floor. The remaining PE GPCIs range from
0.852 (Mississippi) to 1.435 (San Jose, California.)
}

\subsection{Malpractice Expense}{

The Professional Liability Insurance (PLI), which Medicare regulations refer
to as the malpractice GPCI, account for geographic differences in PLI premium
costs. These have the most variation of the three GPCIs, ranging from 0.3
(Minnesota) to 2.5 (Miami).

36 states and territories have a single set of GPCIs. The remaining 14 states
differentiate GPCIs by urban and rural areas. California has the most
differentiation, with 29 different geographic localities.
}

\subsection{Geographic Adjustment Factor (GAF)}{

The three GPCI components can be combined into a composite GPCI (or GAF) by
weighting each by the share of Medicare payments accounted for by the work,
practice cost, and PLI components. GAF indicates how Medicare payments in a
locality differ from the national average (the average being 1.00).

GAFs are not used in the calculation of the Medicare payment schedule amount,
but are a useful measure to illustrate the overall effect of geographic
adjustments under the PFS across Medicare fee schedule areas.
}
}
\examples{
search_gpcis(state = "GA", locality = "01", mac = "10212")

}
\seealso{
Other Physician Fee Schedule Sources: 
\code{\link{search_anesthesia}()},
\code{\link{search_opps}()},
\code{\link{search_payment}()},
\code{\link{search_rvus}()}
}
\concept{Physician Fee Schedule Sources}
