% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_hcpcs}
\alias{search_hcpcs}
\title{Search HCPCS Level II}
\usage{
search_hcpcs(hcpcs = NULL, limit = TRUE, ...)
}
\arguments{
\item{hcpcs}{< \emph{chr} > 5-digit Level II HCPCS code}

\item{limit}{< \emph{lgl} > limit columns returned, default is \code{TRUE}}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search HCPCS Level II
}
\examples{
search_hcpcs(c("A0021", "V5362", "J9264", "G8916"))
}
