% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{search_level_two}
\alias{search_level_two}
\title{HCPCS Level II Codes}
\usage{
search_level_two(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
HCPCS Level II Codes
}
\examples{
search_level_two(hcpcs = c("A0021", "V5362", "J9264", "G8916"))

}
