% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_modifiers}
\alias{search_modifiers}
\title{Level I and II HCPCS Modifiers}
\usage{
search_modifiers(mod_code = NULL, mod_type = NULL, ...)
}
\arguments{
\item{mod_code}{\verb{<chr>} vector of 2-character HCPCS modifiers; default is
\code{NULL}}

\item{mod_type}{\verb{<chr>} Modifier type, one of \code{HCPCS}, \code{CPT},
\code{Anesthesia}, or \verb{Performance Measure}; default is \code{NULL}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
A modifier provides the means to report or indicate that a service or
procedure that has been performed has been altered by some specific
circumstance but not changed in its definition or code.
}
\details{
Modifiers also enable health care professionals to effectively respond to
payment policy requirements established by other entities.
}
\examples{
search_modifiers(mod_code = c("25", "59"))

search_modifiers(mod_type = "CPT")

}
\seealso{
Other HIPAA Code Standards: 
\code{\link{search_adjustments}()},
\code{\link{search_hcpcs}()},
\code{\link{search_pos}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Code Standards}
