% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phys_sched.R
\name{search_opps}
\alias{search_opps}
\title{Outpatient Prospective Payment System (OPPS) Capitations}
\usage{
search_opps(hcpcs_code = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Contains the payment amounts after the application of the OPPS-based payment
caps, except for carrier priced codes. For carrier price codes, the field
only contains the OPPS-based payment caps. Carrier prices cannot exceed the
OPPS-based payment caps.
}
\examples{
head(search_opps())

}
