% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_payment}
\alias{search_payment}
\title{Payment Amount File}
\usage{
search_payment(hcpcs = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{Empty dots}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Payment Amount File
}
\examples{
search_payment(hcpcs    = c("39503", "43116", "33935", "11646"),
               locality = "01",
               mac      = "10212")

}
