% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_plas}
\alias{search_plas}
\title{Proprietary Laboratory Analyses (PLA) Codes
(U codes)
https://www.ama-assn.org/practice-management/cpt/cpt-pla-codes}
\usage{
search_plas(hcpcs_code = NULL, ...)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Proprietary Laboratory Analyses (PLA) Codes
(U codes)
https://www.ama-assn.org/practice-management/cpt/cpt-pla-codes
}
\examples{
search_plas()

}
\seealso{
Other HIPAA Standards: 
\code{\link{search_descriptions}()},
\code{\link{search_hcpcs}()},
\code{\link{search_modifiers}()},
\code{\link{search_pos}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Standards}
