% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_pos}
\alias{search_pos}
\title{Place of Service (POS) Codes}
\usage{
search_pos(pos_code = NULL, pos_type = NULL, ...)
}
\arguments{
\item{pos_code}{\verb{<chr>} vector of 2-character Place of Service codes; default
is \code{NULL}}

\item{pos_type}{\verb{<chr>} Place of Service type, one of \code{Facility} or
\code{Non-Facility}; default is \code{NULL}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Place of Service Codes are two-digit codes placed on health care professional
claims to indicate the setting in which a service was provided. The Centers
for Medicare & Medicaid Services (CMS) maintain POS codes used throughout the
health care industry.
}
\details{
This code set is required for use in the implementation guide adopted as the
national standard for electronic transmission of professional health care
claims under the provisions of the Health Insurance Portability and
Accountability Act of 1996 (HIPAA).

HIPAA directed the Secretary of HHS to adopt national standards for
electronic transactions. These standard transactions require all health plans
and providers to use standard code sets to populate data elements in each
transaction.

The Transaction and Code Set Rule adopted the ASC X12N-837 Health Care Claim:
Professional, volumes 1 and 2, version 4010, as the standard for electronic
submission of professional claims.

This standard names the POS code set currently maintained by CMS as the code
set to be used for describing sites of service in such claims. POS
information is often needed to determine the acceptability of direct billing
of Medicare, Medicaid and private insurance services provided by a given
provider.
\subsection{Code Structure}{

The first digit denotes service category and the second specifies location or
service type.

POS code 11 represents an Office for direct patient services.

For instance, code 11 designates an Office, encompassing
physician offices, clinics, group practices, and standalone facilities
providing direct patient services.

Conversely, place of service code 22
denotes an On Campus-Outpatient Hospital, covering services in a
hospital-based outpatient department where the patient is admitted as an
outpatient.
\itemize{
\item Structure: First digit for category, second for location/type
\item Choose POS based on the majority of services in a specific encounter
\item Modify POS codes if the location changes during a single encounter
}
}

\subsection{Location Types}{

Based on the location of service, POS codes can be grouped into four
categories: Facility, Non-Facility, Telehealth, and Other.
\subsection{Facility Codes}{

Facility POS codes are used to indicate services provided in a facility
setting such as hospitals, nursing homes, or skilled nursing facilities.
These include:
\itemize{
\item Urgent Care Facility (20)
\item Inpatient Hospital (21)
\item Outpatient Hospital (22)
\item Emergency Room-Hospital (23)
\item Ambulatory Surgical Center (24)
\item Skilled Nursing Facility (31)
\item Hospice Facility (32)
}

These codes indicate that the services were provided in a facility that is
owned and operated by a healthcare provider.
}

\subsection{Non-Facility Codes}{

Non-facility POS codes are used for services provided in non-facility
settings such as physician offices or independent clinics. These include:
\itemize{
\item School (03)
\item Office (11)
\item Home (12)
\item Independent Clinic (49)
}

The above codes indicate that the services were provided in a setting not
owned or operated by a healthcare provider.
}

\subsection{Telehealth POS Codes}{

Telehealth place-of-service codes are used to indicate services provided
through telecommunication technology. These include:
\itemize{
\item Telehealth (02)
\item Store and Forward Telemedicine Services (18)
}

This category of POS codes was introduced due to the increasing use of
telehealth services in healthcare, and to differentiate them from traditional
in-person services.
}

\subsection{Other POS Codes}{

In addition to facility-specific POS codes, there are other codes that play a
vital role in accurately describing healthcare encounters. These codes are
designed for specific scenarios, such as visits to retail clinics, public
health clinics, or rural health clinics. Each code serves a distinct purpose
in the healthcare landscape.
\itemize{
\item Homeless Shelter (04)
\item Retail Clinic (17)
\item Rural Health Clinic (72)
}

These codes, including Retail Clinic (17), Home Shelter (04), and Rural
Health Clinic (72), cater to unique healthcare settings, ensuring
comprehensive coverage in billing and reporting. Understanding and applying
these codes appropriately contributes to the precision of healthcare
documentation and facilitates effective reimbursement processes.
}

}
}
\examples{
search_pos(pos_code = c("11", "21"))

search_pos(pos_type = "Facility")

}
\seealso{
Other HIPAA Standards: 
\code{\link{search_descriptions}()},
\code{\link{search_hcpcs}()},
\code{\link{search_modifiers}()},
\code{\link{search_plas}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Standards}
