% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncci.R
\name{search_ptp}
\alias{search_ptp}
\title{Procedure to Procedure (PTP) Edits}
\usage{
search_ptp(column_1 = NULL, column_2 = NULL, mod = NULL, ...)
}
\arguments{
\item{column_1}{vector of HCPCS codes}

\item{column_2}{vector of HCPCS codes}

\item{mod}{1 (Allowed), 0 (Not Allowed), 9 (Not Applicable)}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits
prevent inappropriate payment of services that should not be reported
together. Each edit has a Column One and Column Two HCPCS/CPT code. If a
provider reports the two codes of an edit pair for the same beneficiary on
the same date of service, the Column One code is eligible for payment, but
the Column Two code is denied unless a clinically appropriate NCCI
PTP-associated modifier is also reported.
}
\details{
\href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-procedure-procedure-ptp-edits}{PTP Link}
}
\examples{
search_ptp(column_1 = c("39503", "43116", "33935", "11646"))
}
