% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncci.R
\name{search_ptps}
\alias{search_ptps}
\title{NCCI Procedure to Procedure (PTP) Edits}
\usage{
search_ptps(
  hcpcs_code = NULL,
  ptp_service = c("Practitioner", "Outpatient"),
  ptp_type = NULL,
  ptp_mod = NULL,
  unnest = FALSE,
  ...
)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{ptp_service}{\verb{<chr>} \code{Practitioner} or \code{Outpatient}; default is
\code{Practitioner}}

\item{ptp_type}{\verb{<chr>} \code{Comprehensive} (Column One) or \code{Component} (Column
Two)}

\item{ptp_mod}{\verb{<int>}
\itemize{
\item \code{1}: Allowed: The code-pair is allowed with an NCCI PTP-associated modifier.
\item \code{0}: Not Allowed: An NCCI PTP-associated modifier is not allowed and will not bypass the edit.
\item \code{9}: Not Applicable: The use of NCCI PTP-associated modifiers is not specified.
}}

\item{unnest}{\verb{<lgl>} Unnest the \code{ptp_complements} column, default is \code{FALSE}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits
prevent inappropriate payment of services that should not be reported
together. Each edit has a Column One (Comprehensive) and Column Two
(Component) HCPCS code.
}
\details{
If a provider reports the two codes of an edit pair for the same beneficiary
on the same date of service, the Column One code is eligible for payment, but
the Column Two code is denied unless a clinically appropriate NCCI
PTP-associated modifier is also reported.

The NCCI \href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-procedure-procedure-ptp-edits}{PTP edits}
and
\href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-medically-unlikely-edits}{MUEs}
are usually updated at least quarterly.
\subsection{PTP Modifiers}{

Modifiers that may be used under appropriate clinical circumstances to bypass
an NCCI PTP edit include:
\itemize{
\item Anatomic: E1-E4, FA, F1-F9, TA, T1-T9, LT, RT, LC, LD, RC, LM, RI
\item Global Surgery: 24, 25, 57, 58, 78, 79
\item Other: 27, 59, 91, XE, XS, XP, XU
}

It’s very important that NCCI PTP-associated modifiers only be used when
appropriate. In general, these circumstances relate to separate patient
encounters, separate anatomic sites, or separate specimens.

Most edits involving paired organs or structures (e.g., eyes, ears,
extremities, lungs, kidneys) have NCCI PTP modifier indicators of \code{1} because
the two codes of the code-pair edit may be reported if performed on the
\strong{contralateral} organs or structures.

Most of these code-pairs should not be reported with NCCI PTP-associated
modifiers when performed on the \strong{ipsilateral} organ or structure unless
there is a specific coding rationale to bypass the edit.

The presence of an NCCI PTP edit indicates that the two codes generally can’t
be reported together unless the two corresponding procedures are performed at
two separate patient encounters or two separate anatomic locations.

Similarly, if the two corresponding procedures are performed at the same
patient encounter and in contiguous structures in the same organ or anatomic
region, NCCI PTP-associated modifiers generally shouldn’t be used.

Some Column One/Column Two correct coding edits would never warrant the use
of any of the modifiers associated with the NCCI PTP edits. These code pairs
are assigned a Correct Coding Modifier Indicator (CCMI) of \code{0}.

NCCI PTP modifier indicators of \code{9} indicate that the use of NCCI
PTP-associated modifiers is not specified. This indicator is used for all
code pairs that have a deletion date that is the same as the effective date.
This indicator prevents blank spaces from appearing in the indicator field.

Modifier 59 may be used only if no other appropriate modifier describes the
service. Claim line edits allow use of NCCI PTP-associated \strong{Modifier 91} to
bypass them if one or more of the individual laboratory tests are repeated on
the same date of service.

The repeat testing must be medically reasonable and necessary and can’t be
performed to "confirm initial results; due to testing problems with specimens
and equipment or for any other reason when a normal, one-time, reportable
result is all that is required."
\itemize{
\item \href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-procedure-procedure-ptp-edits}{Procedure to Procedure (PTP) Edits}
}
}
}
\examples{
search_ptps(hcpcs_code = c("39503", "43116", "33935", "11646"))

search_ptps(hcpcs_code = "43116",
            ptp_type = "Component",
            ptp_mod = 0)

}
\seealso{
Other NCCI Edits: 
\code{\link{search_aocs}()},
\code{\link{search_mues}()}
}
\concept{NCCI Edits}
