% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pfs.R
\name{search_rvus}
\alias{search_rvus}
\title{Get HCPCS Codes with RVUs}
\usage{
search_rvus(hcpcs_code = NULL, ...)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Physician Fee Schedule Relative Value File
}
\examples{
search_rvus(hcpcs_code = c("95907", "78140", "32820", "61575"))

}
