% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{single_line_string}
\alias{single_line_string}
\title{Format multiple line character vector to single line}
\usage{
single_line_string(x)
}
\arguments{
\item{x}{character vector}
}
\description{
Format multiple line character vector to single line
}
\keyword{internal}
