% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_addon_edits}
\alias{switch_addon_edits}
\title{NCCI Add-On Code Edit Type Descriptions}
\usage{
switch_addon_edits(x)
}
\arguments{
\item{x}{vector of Add-On Code Edit Types}
}
\value{
vector of Add-On Code Edit Type descriptions
}
\description{
NCCI Add-On Code Edit Type Descriptions
}
\examples{
switch_addon_edits(as.character(1:3))

}
\seealso{
Other NCCI Helpers: 
\code{\link{switch_mue_mai}()},
\code{\link{switch_ptp_mod}()}
}
\concept{NCCI Helpers}
