% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_asc_group}
\alias{switch_asc_group}
\title{HCPCS ASC Group Descriptions}
\usage{
switch_asc_group(x)
}
\arguments{
\item{x}{vector of ASC Group indicators}
}
\value{
vector of ASC Group descriptions
}
\description{
HCPCS ASC Group Descriptions
}
\examples{
dplyr::tibble(
  code = c("YY", 99),
  desc = switch_asc_group(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_betos}()},
\code{\link{switch_coverage}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
