% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_assistant_surgery}
\alias{switch_assistant_surgery}
\title{Assistant Surgery Descriptions}
\usage{
switch_assistant_surgery(x)
}
\arguments{
\item{x}{vector of Assistant Surgery indicators}
}
\value{
vector of Assistant Surgery descriptions
}
\description{
Fee schedule amount equals 16 percent of amount
otherwise applicable for surgical payment.
}
\details{
Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}
}
\examples{
switch_assistant_surgery(c(0:2, "9"))

}
