% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_betos}
\alias{switch_betos}
\title{HCPCS BETOS Descriptions}
\usage{
switch_betos(x)
}
\arguments{
\item{x}{vector of BETOS codes}
}
\value{
vector of BETOS descriptions
}
\description{
HCPCS BETOS Descriptions
}
\examples{
dplyr::tibble(code = c("D1A", "I2A", "M3", "P0"),
              desc = switch_betos(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_coverage}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
