% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_bilateral_surgery}
\alias{switch_bilateral_surgery}
\title{PFS Bilateral Surgery Descriptions}
\usage{
switch_bilateral_surgery(x)
}
\arguments{
\item{x}{vector of Bilateral Surgery indicators}
}
\value{
vector of Bilateral Surgery descriptions
}
\description{
Modifier 50: Indicates services subject to payment adjustment
}
\examples{
switch_bilateral_surgery(c(0:3, "9"))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
