% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_co_surgeon}
\alias{switch_co_surgeon}
\title{PFS Co-Surgeon Descriptions}
\usage{
switch_co_surgeon(x)
}
\arguments{
\item{x}{vector of Co-Surgeon indicators}
}
\value{
vector of Co-Surgeon descriptions
}
\description{
Modifier 62: Services for which two surgeons,
each in a different specialty, may be paid.
}
\details{
Reimbursement is 62.5\% of MPFS allowed amount.
}
\examples{
switch_co_surgeon(as.character(0:9))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
