% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_coverage}
\alias{switch_coverage}
\title{HCPCS Coverage Descriptions}
\usage{
switch_coverage(x)
}
\arguments{
\item{x}{vector of Coverage indicators}
}
\value{
vector of Coverage descriptions
}
\description{
HCPCS Coverage Descriptions
}
\examples{
dplyr::tibble(
  code = c("C", "D", "I", "M", "S"),
  desc = switch_coverage(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_betos}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
