% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_diagnostic_imaging}
\alias{switch_diagnostic_imaging}
\title{PFS Diagnostic Imaging Descriptions}
\usage{
switch_diagnostic_imaging(x)
}
\arguments{
\item{x}{vector of Diagnostic Imaging indicators}
}
\value{
vector of Diagnostic Imaging descriptions
}
\description{
Identifies the applicable Diagnostic Service family for
HCPCS codes with a Multiple Procedure indicator of 4.
}
\examples{
dplyr::tibble(
  code = fuimus::pad_number(c(1:11, 88, 99)),
  desc = switch_diagnostic_imaging(code))

}
