% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_global_days}
\alias{switch_global_days}
\title{PFS Global Days Descriptions}
\usage{
switch_global_days(x)
}
\arguments{
\item{x}{vector of Global Days indicators}
}
\value{
vector of Global Days descriptions
}
\description{
PFS Global Days Descriptions
}
\examples{
x <- c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ")
switch_global_days(x)

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
