% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_lab_certification}
\alias{switch_lab_certification}
\title{HCPCS Lab Certification Descriptions}
\usage{
switch_lab_certification(x)
}
\arguments{
\item{x}{vector of Lab Certification indicators}
}
\value{
vector of Lab Certification descriptions
}
\description{
HCPCS Lab Certification Descriptions
}
\examples{
dplyr::tibble(
  code = c("010", 100, 110, 115, 120, "110, 120, 130, 400"),
  desc = switch_lab_certification(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_betos}()},
\code{\link{switch_coverage}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
