% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_multiple_pricing}
\alias{switch_multiple_pricing}
\title{HCPCS Multiple Pricing Descriptions}
\usage{
switch_multiple_pricing(x)
}
\arguments{
\item{x}{vector of Multiple Pricing indicators}
}
\value{
vector of Multiple Pricing descriptions
}
\description{
HCPCS Multiple Pricing Descriptions
}
\examples{
dplyr::tibble(code = c(9, LETTERS[1:7]))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_betos}()},
\code{\link{switch_coverage}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
