% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_multiple_procedure}
\alias{switch_multiple_procedure}
\title{PFS Multiple Procedure Descriptions}
\usage{
switch_multiple_procedure(x)
}
\arguments{
\item{x}{vector of Multiple Procedure indicators}
}
\value{
vector of Multiple Procedure descriptions
}
\description{
Modifier 51: Indicates applicable payment adjustment rule for multiple
procedures
}
\details{
Multiple Surgery Rules
100\% of MPFS amount is allowed for highest valued surgical procedure and 50\%
for additional surgical procedures (with a multiple surgery indicator of "2")
performed same day. Modifier 51 will be appended to identify reduced
services.
}
\examples{
switch_multiple_procedure(as.character(0:9))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
