% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_fee.R
\name{switch_opps_indicator}
\alias{switch_opps_indicator}
\title{OPPS Indicator Descriptions}
\usage{
switch_opps_indicator(x)
}
\arguments{
\item{x}{vector of OPPS indicators}
}
\value{
vector of OPPS indicator descriptions
}
\description{
OPPS Indicator Descriptions
}
\examples{
switch_opps_indicator(c("1", "9"))

}
