% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_pctc_indicator}
\alias{switch_pctc_indicator}
\title{PFS PC/TC Indicator Descriptions}
\usage{
switch_pctc_indicator(x)
}
\arguments{
\item{x}{vector of PC/TC indicators}
}
\value{
vector of PC/TC indicator descriptions
}
\description{
PFS PC/TC Indicator Descriptions
}
\examples{
switch_pctc_indicator(as.character(0:9))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
