% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_pricing}
\alias{switch_pricing}
\title{HCPCS Pricing Descriptions}
\usage{
switch_pricing(x)
}
\arguments{
\item{x}{vector of Pricing indicators}
}
\value{
vector of Pricing descriptions
}
\description{
HCPCS Pricing Descriptions
}
\examples{
dplyr::tibble(
  code = c("00", 11:13, 21:22, 31:46, 51:57, 99),
  desc = switch_pricing(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_betos}()},
\code{\link{switch_coverage}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_type_of_service}()}
}
\concept{HCPCS II Helpers}
