% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_ptp_mod}
\alias{switch_ptp_mod}
\title{NCCI PTP CCMI Descriptions}
\usage{
switch_ptp_mod(x)
}
\arguments{
\item{x}{vector of PTP CCMIs}
}
\value{
vector of PTP CCMI descriptions
}
\description{
NCCI PTP CCMI Descriptions
}
\examples{
switch_ptp_mod(as.character(0:9))

}
\seealso{
Other NCCI Helpers: 
\code{\link{switch_addon_edits}()},
\code{\link{switch_mue_mai}()}
}
\concept{NCCI Helpers}
