% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_fee.R
\name{switch_status_code}
\alias{switch_status_code}
\title{Status Code Descriptions}
\usage{
switch_status_code(x)
}
\arguments{
\item{x}{vector of PC/TC indicators}
}
\value{
vector of PC/TC indicator descriptions
}
\description{
Status Code Descriptions
}
\examples{
switch_status_code(LETTERS)

}
\keyword{internal}
