% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_status_description}
\alias{switch_status_description}
\title{Status Code Descriptions}
\usage{
switch_status_description(x)
}
\arguments{
\item{x}{vector of status codes}
}
\value{
vector of status code descriptions
}
\description{
Status Code Descriptions
}
\examples{
dplyr::tibble(
  code = LETTERS,
  name = switch_status_name(LETTERS),
  desc = switch_status_description(LETTERS)
)

}
