% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_status_name}
\alias{switch_status_name}
\title{PFS Status Code Names}
\usage{
switch_status_name(x)
}
\arguments{
\item{x}{vector of status codes}
}
\value{
vector of status code names
}
\description{
PFS Status Code Names
}
\examples{
dplyr::tibble(
  code = LETTERS,
  name = switch_status_name(LETTERS),
  desc = switch_status_description(LETTERS)
)

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
