% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_supervision}
\alias{switch_supervision}
\title{PFS Physician Supervision Descriptions}
\usage{
switch_supervision(x)
}
\arguments{
\item{x}{vector of Physician Supervision indicators}
}
\value{
vector of Physician Supervision descriptions
}
\description{
This field is for use in post payment review.
}
\examples{
dplyr::tibble(
  code = c(fuimus::pad_number(
  c(1:6, 9, 21:22, 66, 77)), "6A", "7A"),
  desc = switch_supervision(code))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
