% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_team_surgery}
\alias{switch_team_surgery}
\title{PFS Team Surgery Descriptions}
\usage{
switch_team_surgery(x)
}
\arguments{
\item{x}{vector of Team Surgery indicators}
}
\value{
vector of Team Surgery descriptions
}
\description{
Modifier 66: Services for which Team Surgeons may be paid
}
\examples{
switch_team_surgery(as.character(0:9))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()}
}
\concept{PFS Helpers}
