% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_type_of_service}
\alias{switch_type_of_service}
\title{HCPCS Type of Service Descriptions}
\usage{
switch_type_of_service(x)
}
\arguments{
\item{x}{vector of Type of Service codes}
}
\value{
vector of Type of Service descriptions
}
\description{
HCPCS Type of Service Descriptions
}
\examples{
dplyr::tibble(code = c(0:9, LETTERS),
              desc = switch_type_of_service(code))

}
\seealso{
Other HCPCS II Helpers: 
\code{\link{switch_action_code}()},
\code{\link{switch_asc_group}()},
\code{\link{switch_betos}()},
\code{\link{switch_coverage}()},
\code{\link{switch_lab_certification}()},
\code{\link{switch_multiple_pricing}()},
\code{\link{switch_pricing}()}
}
\concept{HCPCS II Helpers}
