% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_amounts}
\alias{calc_amounts}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
calc_amounts(wrvu, fprvu, nprvu, mrvu, wgpci, pgpci, mgpci, cf)
}
\arguments{
\item{wrvu}{Work RVUs}

\item{fprvu}{Facility Practice Expense RVUs}

\item{nprvu}{Non-Facility Practice Expense RVUs}

\item{mrvu}{Malpractice RVUs}

\item{wgpci}{Work GPCI}

\item{pgpci}{Practice Expense GPCI}

\item{mgpci}{Malpractice GPCI}

\item{cf}{Conversion Factor}
}
\value{
Facility & Non-Facility Participating, Non-Participating & Limiting Charge Amounts
}
\description{
Calculate Physician Fee Schedule Payment Amounts
}
\examples{
calc_amounts(wrvu = 6.26,
             nprvu = 7.92,
             fprvu = 4.36,
             mrvu = 0.99,
             wgpci = 1.0,
             pgpci = 0.883,
             mgpci = 1.125,
             cf = 32.7442)
}
