% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_amounts}
\alias{calc_amounts}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
calc_amounts(wrvu, prvu, mrvu, wgpci, pgpci, mgpci, cf)
}
\arguments{
\item{wrvu}{numeric}

\item{prvu}{numeric}

\item{mrvu}{numeric}

\item{wgpci}{numeric}

\item{pgpci}{numeric}

\item{mgpci}{numeric}

\item{cf}{numeric}
}
\value{
description
}
\description{
Calculate Physician Fee Schedule Payment Amounts
}
\examples{
calc_amounts(wrvu  = 108.91,
             prvu  = 35.14,
             mrvu  = 26.95,
             wgpci = 1.0,
             pgpci = 0.997,
             mgpci = 1.128,
             cf    = 32.7442)
}
