% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{calc_amounts_df}
\alias{calc_amounts_df}
\title{Calculate Physician Fee Schedule Payment Amounts}
\usage{
calc_amounts_df(hcpcs, state = NULL, locality = NULL, mac = NULL)
}
\arguments{
\item{hcpcs}{numeric}

\item{state}{numeric}

\item{locality}{numeric}

\item{mac}{numeric}
}
\value{
description
}
\description{
Calculate Physician Fee Schedule Payment Amounts
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
calc_amounts_df(hcpcs = "11646",
                state = "GA",
                locality = "99",
                mac = "10212")
\dontshow{\}) # examplesIf}
}
