% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-labeling.R
\name{case_status}
\alias{case_status}
\title{Replace Status Codes with Labels, Add Descriptions}
\usage{
case_status(df, col, desc = FALSE)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Status Codes}

\item{desc}{add Descriptions}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{status_desc} column
}
\description{
Replace Status Codes with Labels, Add Descriptions
}
\examples{
dplyr::tibble(status = LETTERS) |>
case_status(status, desc = TRUE)
}
