% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level2.R
\name{case_coverage}
\alias{case_coverage}
\title{Replace HCPCS Level II Coverage Indicators with Descriptions}
\usage{
case_coverage(df, col)
}
\arguments{
\item{df}{data frame}

\item{col}{column of Coverage indicators}
}
\value{
A \link[tibble:tibble-package]{tibble} with a \code{cov_description} column
}
\description{
Replace HCPCS Level II Coverage Indicators with Descriptions
}
\examples{
dplyr::tibble(cov = c("C", "D", "I", "M", "S")) |> case_coverage(cov)
}
