% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{construct_regex}
\alias{construct_regex}
\title{Construct regex pattern for HCPCS codes}
\usage{
construct_regex(x)
}
\arguments{
\item{x}{\verb{<chr>} vector}
}
\value{
\verb{<chr>} vector
}
\description{
Construct regex pattern for HCPCS codes
}
\examples{
construct_regex(search_descriptions()$hcpcs_code)

}
\keyword{internal}
