% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{count_days}
\alias{count_days}
\title{Count days between two dates}
\usage{
count_days(df, start, end, name)
}
\arguments{
\item{df}{data frame}

\item{start}{date column}

\item{end}{date column}

\item{name}{name of output column}
}
\value{
A \link[tibble:tibble-package]{tibble}
}
\description{
Count days between two dates
}
\examples{
dplyr::tibble(id = 1:3,
              dos = as.Date(c("2021-04-18", "2021-11-18", "2022-02-18")),
              signed = as.Date("2022-02-18")) |>
              count_days(start = dos, end = signed, provider_lag)
}
\keyword{internal}
