% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{descriptors}
\alias{descriptors}
\title{2023 CPT Descriptors (Clinician & Consumer-Friendly)}
\usage{
descriptors(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{5-digit Level I, Category I HCPCS code}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2023 CPT Descriptors (Clinician & Consumer-Friendly)
}
\examples{
descriptors(c("39503", "43116", "33935", "11646"))
}
