% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncci.R
\name{get_ptp_edits}
\alias{get_ptp_edits}
\title{Procedure to Procedure (PTP) Edits}
\usage{
get_ptp_edits(
  hcpcs_code = NULL,
  ptp_type = NULL,
  ptp_edit_mod = NULL,
  current = TRUE,
  ...
)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{ptp_type}{\verb{<chr>} \code{comprehensive} (Column One code) or \code{component}
(Column Two code)}

\item{ptp_edit_mod}{\verb{<int>}
\itemize{
\item \code{1}: Allowed
\item \code{0}: Not Allowed
\item \code{9}: Not Applicable
}}

\item{current}{\verb{<lgl>} return only current edits, default is \code{TRUE}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
National Correct Coding Initiative (NCCI) Procedure-to-Procedure (PTP) edits
prevent inappropriate payment of services that should not be reported
together. Each edit has a Column One and Column Two HCPCS/CPT code.
}
\details{
If a provider reports the two codes of an edit pair for the same beneficiary
on the same date of service, the Column One code is eligible for payment, but
the Column Two code is denied unless a clinically appropriate NCCI
PTP-associated modifier is also reported.

The NCCI \href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-procedure-procedure-ptp-edits}{PTP edits} and \href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-medically-unlikely-edits}{MUEs} are usually updated at least quarterly.
\subsection{NCCI Modifiers}{

Modifiers that may be used under appropriate clinical circumstances to bypass
an NCCI PTP edit include:
\itemize{
\item Anatomic: E1-E4, FA, F1-F9, TA, T1-T9, LT, RT, LC, LD, RC, LM, RI
\item Global Surgery: 24, 25, 57, 58, 78, 79
\item Other: 27, 59, 91, XE, XS, XP, XU
}

It’s very important that NCCI PTP-associated modifiers only be used when
appropriate. In general, these circumstances relate to separate patient
encounters, separate anatomic sites, or separate specimens. (See subsequent
discussion of modifiers in this section.)

Most edits involving paired organs or structures (e.g., eyes, ears,
extremities, lungs, kidneys) have NCCI PTP modifier indicators of \code{1} because
the two codes of the code pair edit may be reported if performed on the
contralateral organs or structures.

Most of these code pairs should not be reported with NCCI PTP-associated
modifiers when performed on the ipsilateral organ or structure unless there
is a specific coding rationale to bypass the edit.

The presence of an NCCI PTP edit indicates that the two codes generally can’t
be reported together unless the two corresponding procedures are performed at
two separate patient encounters or two separate anatomic locations.

Similarly, if the two corresponding procedures are performed at the same
patient encounter and in contiguous structures in the same organ or anatomic
region, NCCI PTP-associated modifiers generally shouldn’t be used.

Modifier 59 (\href{https://www.cms.gov/files/document/mln1783722-proper-use-modifiers-59-xe-xp-xs-and-xu.pdf}{MLN 1783722})
may be used only if no other appropriate modifier describes the service. The
article provides more information on the appropriate use of the 59 modifier.

Some Column One/Column Two correct coding edits would ever warrant the use of
any of the modifiers associated with the NCCI PTP edits. These code pairs are
assigned a correct coding modifier indicator (CCMI) of \code{0}.

Claim line edits allow use of NCCI PTP-associated \strong{Modifier 91} to bypass
them if one or more of the individual laboratory tests are repeated on the
same date of service.

The repeat testing must be medically reasonable and necessary and can’t be
performed to "confirm initial results; due to testing problems with specimens
and equipment or for any other reason when a normal, one-time, reportable
result is all that is required."
\itemize{
\item \href{https://www.cms.gov/medicare/coding-billing/national-correct-coding-initiative-ncci-edits/medicare-ncci-procedure-procedure-ptp-edits}{Procedure to Procedure (PTP) Edits}
}
}
}
\examples{
get_ptp_edits(hcpcs_code = c("39503", "43116", "33935", "11646"))

get_ptp_edits(hcpcs_code = "43116",
              ptp_type     = "component",
              ptp_edit_mod = 0)

}
