% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_charge.R
\name{gpci}
\alias{gpci}
\title{GPCIs 2024}
\usage{
gpci()
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
GPCIs 2024
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
gpci()
\dontshow{\}) # examplesIf}
}
