% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{gpci}
\alias{gpci}
\title{2024 Geographic Practice Cost Indices}
\usage{
gpci(mac = NULL, state = NULL, locality = NULL)
}
\arguments{
\item{mac}{description}

\item{state}{description}

\item{locality}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2024 Geographic Practice Cost Indices
}
\examples{
gpci(state    = "GA",
     locality = "01",
     mac      = "10212") |>
dplyr::glimpse()
}
