% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{gpci}
\alias{gpci}
\title{2024 Geographic Practice Cost Indices}
\usage{
gpci(mac = NULL, state = NULL, locality = NULL, ...)
}
\arguments{
\item{mac}{5-digit Medicare Administrative Contractor (MAC) code}

\item{state}{2-character State abbreviation}

\item{locality}{2-digit locality ID}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2024 Geographic Practice Cost Indices
}
\examples{
gpci(state    = "GA",
     locality = "01",
     mac      = "10212") |>
dplyr::glimpse()
}
