% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{hcpcs_search}
\alias{hcpcs_search}
\title{Look up Information about HCPCS Codes}
\usage{
hcpcs_search(hcpcs, state = NULL, locality = NULL, mac = NULL)
}
\arguments{
\item{hcpcs}{< \emph{character} > 5-character HCPCS Code}

\item{state}{< \emph{character} > 2-character state abbreviation}

\item{locality}{< \emph{character} > 2-character locality id}

\item{mac}{< \emph{character} > 5-character MAC id code}
}
\value{
A \link[tibble:tibble-package]{tibble} with the columns:\tabular{ll}{
   \strong{Column} \tab \strong{Description} \cr
   \code{hcpcs} \tab 5-character HCPCS Code \cr
   \code{hcpcs_type} \tab HCPCS Type \cr
   \code{description} \tab HCPCS Description \cr
   \code{cons_desc} \tab HCPCS Level I Consumer Description \cr
   \code{clin_descs} \tab HCPCS Level I Clinician Descriptions \cr
   \code{mod} \tab Modifier \cr
   \code{status} \tab Status Code \cr
   \code{mac} \tab Medicare Administrative Contractor ID \cr
   \code{state} \tab State Abbreviation \cr
   \code{locality} \tab State Abbreviation \cr
   \code{area} \tab State Abbreviation \cr
   \code{counties} \tab State Abbreviation \cr
   \code{two_macs} \tab State Abbreviation \cr
   \code{wgpci} \tab State Abbreviation \cr
   \code{pgpci} \tab State Abbreviation \cr
}
}
\description{
Look up Information about HCPCS Codes
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
hcpcs_search(hcpcs = "V5299",
             state = "GA",
             locality = "99",
             mac = "10212")

hcpcs_search(hcpcs = c("39503", "43116", "33935", "11646"), state = "GA")
\dontshow{\}) # examplesIf}
}
