% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/icd10.R
\name{icd10cm}
\alias{icd10cm}
\title{2024 National Physician Fee Schedule Relative Value File}
\usage{
icd10cm(code = NULL)
}
\arguments{
\item{code}{ICD-10-CM code}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2024 National Physician Fee Schedule Relative Value File
}
\examples{
icd10cm(c("H00.019", "D50.1", "C4A.70", "Z20.818")) |> dplyr::glimpse()
}
