% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_aoc_type}
\alias{is_aoc_type}
\title{What type of Add-On Code is a HCPCS Code?}
\usage{
is_aoc_type(hcpcs)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of HCPCS codes}
}
\value{
\verb{<list>} of three vectors: \code{primary}, \code{addon}, \code{both}
}
\description{
What type of Add-On Code is a HCPCS Code?
}
\examples{
is_aoc_type(c("11646", "0074T"))

c("22633", "22630", "22532", "77001", "88334", "0715T", "64727") |>
is_aoc_type()
}
