% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_category_I}
\alias{is_category_I}
\title{Check if code is HCPCS Level 1 Category I (CPT)}
\usage{
is_category_I(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{\verb{<chr>} string of \code{length(1)}}

\item{arg}{\verb{<chr>} function argument name in the current function}

\item{call}{\verb{<environment>} environment the function is called from}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, \code{FALSE} otherwise
}
\description{
Check if code is HCPCS Level 1 Category I (CPT)
}
\examples{
is_category_I("11646")

is_category_I("1164F")
}
