% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_category_III}
\alias{is_category_III}
\title{Check if code is HCPCS Level 1 Category III}
\usage{
is_category_III(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{string}

\item{arg}{string}

\item{call}{string}
}
\value{
boolean
}
\description{
Category III: a temporary set of codes for emerging technologies,
services, procedures, and service paradigms.
}
\examples{
is_category_III("0074T")

try(is_category_III("11646"))
}
