% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{is_valid_length}
\alias{is_valid_length}
\title{Check if string is valid length of a HCPCS code}
\usage{
is_valid_length(x, arg = rlang::caller_arg(x), call = rlang::caller_env())
}
\arguments{
\item{x}{\verb{<chr>} string of \code{length(1)}}

\item{arg}{\verb{<chr>} function argument name in the current function}

\item{call}{\verb{<environment>} environment the function is called from}
}
\value{
\verb{<lgl>} \code{TRUE} if valid, \code{FALSE} otherwise
}
\description{
Check if string is valid length of a HCPCS code
}
\examples{
is_valid_length("11646")

try(is_valid_length("1164"))

}
