% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{lcd}
\alias{lcd}
\title{LCD Download Database}
\usage{
lcd()
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Last Updated 2023-04-27
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
lcd()
\dontshow{\}) # examplesIf}
}
