% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lcd_ncd.R
\name{lcd}
\alias{lcd}
\title{LCD Download Database}
\usage{
lcd(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{vector of HCPCS codes}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Last Updated 2023-04-27
}
\examples{
lcd(hcpcs = c("A4555", "E0766"))
}
