% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{level2}
\alias{level2}
\title{2024 Healthcare Common Procedure Coding System (HCPCS)}
\usage{
level2(hcpcs = NULL, limit_cols = TRUE, ...)
}
\arguments{
\item{hcpcs}{5-digit Level II HCPCS code}

\item{limit_cols}{limit the number of columns returned}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2024 Healthcare Common Procedure Coding System (HCPCS)
}
\examples{
level2(c("A0021", "V5362", "J9264", "G8916")) |> dplyr::glimpse()
}
