% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{limiting_charge}
\alias{limiting_charge}
\title{Calculate Limiting Charge}
\usage{
limiting_charge(par_amount)
}
\arguments{
\item{par_amount}{< \emph{double} > Participating Amount}
}
\value{
numeric vector of Limiting Charge Amount
}
\description{
The Medicare limiting charge is set by law at 115\% of the payment amount
for the service furnished by the Non-participating physician.
}
\details{
However, the law sets the payment amount for Non-participating physicians
at 95\% of the payment amount for Participating physicians (i.e., the fee
schedule amount).

Calculating 95\% of 115\% of an amount is equivalent to multiplying the
amount by a factor of 1.0925 (or 109.25\%).

Therefore, to calculate the Medicare limiting charge for a physician
service for a locality, multiply the fee schedule amount by a factor
of 1.0925. The result is the Medicare limiting charge for that service for
that locality to which the fee schedule amount applies.
}
\examples{
limiting_charge(26.35)
}
