% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculations.R
\name{limiting_charge}
\alias{limiting_charge}
\title{Calculate Limiting Charge}
\usage{
limiting_charge(participating_fee)
}
\arguments{
\item{participating_fee}{\verb{<dbl>} Participating Fee}
}
\value{
\verb{<dbl>} vector of the Limiting Charge Amount
}
\description{
The Medicare limiting charge is set by law at 115\% of the payment amount
for the service furnished by the Non-participating physician.
}
\details{
However, the law sets the payment amount for Non-participating physicians
at 95\% of the payment amount for Participating physicians (i.e., the fee
schedule amount).

Calculating 95\% of 115\% of an amount is equivalent to multiplying the
amount by a factor of 1.0925 (or 109.25\%).

Therefore, to calculate the limiting charge, \eqn{x_{limit}} for a physician
service for a locality, multiply the fee schedule amount by a factor
of 1.0925:

\deqn{x_{limit} = \dfrac{\sum_{i = 0}^t
        \sum_{j = 0}^\infty c_i f_{j - i}}{\sum_{i = 0} c_i}}
}
\examples{
limiting_charge(26.35)

}
