% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{ncd}
\alias{ncd}
\title{NCD Download Database}
\usage{
ncd(coverage = NULL)
}
\arguments{
\item{coverage}{Coverage level; \code{"Full"}, \code{"Restricted"}, \code{"None"}, \code{"Unknown"}}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Last Updated 2022-12-08
}
\examples{
ncd(coverage = "Unknown")
}
