% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coverage_determinations.R
\name{ncd}
\alias{ncd}
\title{National Coverage Determinations}
\usage{
ncd(coverage = NULL, ...)
}
\arguments{
\item{coverage}{Coverage level; \code{"Full"}, \code{"Restricted"}, \code{"None"}, \code{"Unknown"}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Database Last Updated 2022-12-08
}
\examples{
ncd(coverage = "Unknown")

}
