% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{opps}
\alias{opps}
\title{OPPSCAP}
\usage{
opps(hcpcs = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs}{description}

\item{mac}{description}

\item{locality}{description}

\item{...}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
Contains the payment amounts after the application of the OPPS-based payment
caps, except for carrier priced codes. For carrier price codes, the field
only contains the OPPS-based payment caps. Carrier prices cannot exceed the
OPPS-based payment caps.
}
\examples{
opps(hcpcs    = c("70170", "71550", "0689T", "75898"),
     mac      = "01112",
     locality = "05")
}
