% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{opps}
\alias{opps}
\title{OPPSCAP}
\usage{
opps(hcpcs = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs}{\emph{\if{html}{\out{<chr>}}} vector of 5-digit HCPCS codes}

\item{mac}{\emph{\if{html}{\out{<chr>}}} vector of 5-digit Medicare Administrative Contractor (MAC) codes}

\item{locality}{\emph{\if{html}{\out{<chr>}}} vector of 2-digit locality IDs}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Contains the payment amounts after the application of the OPPS-based payment
caps, except for carrier priced codes. For carrier price codes, the field
only contains the OPPS-based payment caps. Carrier prices cannot exceed the
OPPS-based payment caps.
}
\examples{
opps(hcpcs    = c("70170", "71550", "0689T", "75898"),
     mac      = "01112",
     locality = "05")
}
