% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{payment}
\alias{payment}
\title{2024 Physician Fee Schedule Payment Amount File}
\usage{
payment(hcpcs = NULL, mac = NULL, locality = NULL)
}
\arguments{
\item{hcpcs}{description}

\item{mac}{description}

\item{locality}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2024 Physician Fee Schedule Payment Amount File
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
payment(hcpcs = "11646", locality = "01", mac = "10212")
\dontshow{\}) # examplesIf}
}
