% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limiting_charge.R
\name{payment}
\alias{payment}
\title{PFS Payment Amount 2024}
\usage{
payment()
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
PFS Payment Amount 2024
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
payment()
\dontshow{\}) # examplesIf}
}
