% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rvu}
\alias{rvu}
\title{2024 National Physician Fee Schedule Relative Value File}
\usage{
rvu(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{description}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
2024 National Physician Fee Schedule Relative Value File
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rvu(hcpcs = "11646")
\dontshow{\}) # examplesIf}
}
