% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rvu}
\alias{rvu}
\title{2024 National Physician Fee Schedule Relative Value File}
\usage{
rvu(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{5-digit HCPCS code}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
2024 National Physician Fee Schedule Relative Value File
}
\examples{
rvu(c("A0021", "V5362", "J9264", "G8916")) |>
dplyr::glimpse()
}
