% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{rvu}
\alias{rvu}
\title{PFS RVU 2024}
\usage{
rvu()
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
PFS RVU 2024
}
\examples{
\dontshow{if (interactive()) (if (getRversion() >= "3.4") withAutoprint else force)(\{ # examplesIf}
rvu()
\dontshow{\}) # examplesIf}
}
