% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjustment.R
\name{search_adjustments}
\alias{search_adjustments}
\title{Search Adjustment Codes}
\usage{
search_adjustments(
  df = NULL,
  col = NULL,
  type = c("all", "carc", "rarc"),
  action = c("review", "join"),
  ...
)
}
\arguments{
\item{df}{< \emph{df} > data.frame}

\item{col}{< \emph{chr} > column of Adjustment codes to match on}

\item{type}{< \emph{chr} > type of Adjustment code; \code{all} (default), \code{carc}, \code{rarc}}

\item{action}{< \emph{chr} > action to take; \code{review} (default), \code{join}}

\item{...}{Empty}
}
\value{
a \code{\link[dplyr:reexports]{dplyr::tibble()}}
}
\description{
CARC and RARC Codes
}
\details{
Claim Adjustment Reason Codes:

\emph{X12 External Code Source 139}

These codes describe why a claim or service line was paid differently
than it was billed and generally assign responsibility for the
adjustment amounts.

The Claim Adjustment \emph{Group Codes} are internal to the X12 standard.
The format is always two alpha characters:
\itemize{
\item \strong{CO}: Contractual Obligations
\item \strong{CR}: Corrections and Reversals
\item \strong{OA}: Other Adjustments
\item \strong{PI}: Payer Initiated Reductions
\item \strong{PR}: Patient Responsibility
}

Remittance Advice Remark Codes:

\emph{X12 External Code Source 411}

Remittance Advice Remark Codes (RARCs) are used to provide additional
explanation for an adjustment already described by a Claim Adjustment Reason
Code (CARC) or to convey information about remittance processing.

There are two types of RARCs: \strong{Supplemental} and \strong{Informational}:

The majority of RARCs are \emph{supplemental} and, as such, are generally referred
to as RARCs without further distinction. Supplemental RARCs provide additional
explanation for an adjustment already described by a CARC.

The second type is \emph{informational}; these are all prefaced with \verb{Alert:} and
are referred to as Alerts. They are used to convey information about
remittance processing and are \emph{never} related to a specific adjustment or CARC.
}
\examples{
dplyr::tibble(code = c("CO-253", "OA-23", "PI-185")) |>
search_adjustments(col    = code,
                   type   = "carc",
                   action = "join")
}
