% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ncci.R
\name{search_aocs}
\alias{search_aocs}
\title{NCCI Add-On Code Edits}
\usage{
search_aocs(
  hcpcs_code = NULL,
  aoc_type = NULL,
  aoc_edit = NULL,
  unnest = FALSE,
  ...
)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{aoc_type}{\verb{<chr>} AOC type; \code{Add-On} or \code{Primary}}

\item{aoc_edit}{\verb{<int>} AOC edit type:
\itemize{
\item \code{1}: Only Paid if Primary is Paid. Payment Eligible if Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.
\item \code{2}: Some Specific Primaries. Payment Eligible if, as Determined by MAC, Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.
\item \code{3}: No Specific Primary Codes. Payment Eligible if, as Determined by MAC, Primary Payment Eligible to Same Practitioner for Same Patient on Same DOS.
}}

\item{unnest}{\verb{<lgl>} Unnest the \code{aoc_complements} column, default is \code{FALSE}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Medicare NCCI Add-On Code Edits
}
\details{
An AOC is a HCPCS/CPT code that describes a service that, with rare
exception, is performed in conjunction with another primary service by the
same practitioner. An AOC is rarely eligible for payment if it is the only
procedure reported by a practitioner.

Some CPT codes are identified as Add-on Codes (AOCs), which describe a
service that can only be reported in addition to a primary procedure.

For specific primary codes, AOCs shall not be reported as a supplemental
service for other HCPCS/CPT codes not listed as a primary code.

AOCs permit the reporting of significant supplemental services commonly
performed in addition to the primary procedure. By contrast, incidental
services that are necessary to accomplish the primary procedure (e.g., lysis
of adhesions in the course of an open cholecystectomy) are not separately
reportable with an AOC. Similarly, complications inherent in an invasive
procedure occurring during the procedure are not separately reportable. For
example, control of bleeding during an invasive procedure is considered part
of the procedure and is not separately reportable.

Although the AOC and primary code are normally reported for the same date of
service, there are unusual circumstances where the two services may be
reported for different dates of service (e.g., 99291 and 99292).
\subsection{AOC Edit Types}{
\itemize{
\item Type 1: CPT Professional or HCPCS files define all acceptable primary codes. MACs should not allow other primary codes with Type 1 AOCs.
\item Type 2: CPT Professional and HCPCS files do not define any primary codes. MACs should develop their own lists of acceptable primary codes.
\item Type 3: CPT Professional or HCPCS files define some, but not all, acceptable primary codes. MACs should allow the listed primary codes for these AOCs but may develop their own lists of additional acceptable primary codes.
}
}

\subsection{PTP Edits In general, NCCI PTP edits do not include edits with most AOCs}{

because edits related to the primary procedure(s) are adequate to prevent
inappropriate payment for an add-on coded procedure (i.e., if an edit
prevents payment of the primary procedure code, the AOC shall not be paid).
}
}
\note{
Version: 2024-04-01
\itemize{
\item \href{https://www.cms.gov/ncci-medicare/medicare-ncci-add-code-edits}{NCCI Add-On Code Edits}
}
}
\examples{
search_aocs(hcpcs_code = "22532", aoc_type = "Primary")

search_aocs(hcpcs_code = c("22630", "77001", "88334", "64727"))

search_aocs(hcpcs_code = "33935", unnest = TRUE)

}
\seealso{
Other NCCI Edits: 
\code{\link{search_mues}()},
\code{\link{search_ptps}()}
}
\concept{NCCI Edits}
