% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_cpt}
\alias{search_cpt}
\title{Search HCPCS Level I (CPT) Codes}
\usage{
search_cpt(hcpcs = NULL)
}
\arguments{
\item{hcpcs}{< \emph{chr} > 5-digit CPT codes}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search HCPCS Level I (CPT) Codes
}
\examples{
search_cpt(c("39503", "43116", "33935", "11646"))
}
