% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_descriptions}
\alias{search_descriptions}
\title{Describe HCPCS Codes}
\usage{
search_descriptions(
  hcpcs_code = NULL,
  hcpcs_desc_type = "All",
  column = "hcpcs_code",
  ...
)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{hcpcs_desc_type}{\verb{<chr>} vector of code description types; \code{All} (default),
\code{Short}, \code{Long}, \code{Medium}, \code{Medical}, \code{Consumer}, \code{Clinician}, \verb{Proprietary Name}}

\item{column}{If \code{hcpcs_code} is a \link{data.frame} or a
\link[tibble:tibble-package]{tibble}, this is the quoted name of the column
containing HCPCS codes; default is \code{"hcpcs_code"}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Describe HCPCS Codes
}
\examples{
search_descriptions(hcpcs_code = c("39503", "43116", "33935", "11646"))

search_descriptions(
  hcpcs_code = dplyr::tibble(code = c("A0021", "V5362", "J9264", "G8916")),
  column = "code")

dplyr::tibble(code = c("A0021", "V5362", "J9264", "G8916")) |>
  search_descriptions(column = "code")

}
\seealso{
Other HIPAA Standards: 
\code{\link{search_hcpcs}()},
\code{\link{search_modifiers}()},
\code{\link{search_pos}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Standards}
