% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fee_schedule.R
\name{search_fee_schedule}
\alias{search_fee_schedule}
\title{Search Physician Fee Schedule}
\usage{
search_fee_schedule(hcpcs, state = NULL, locality = NULL, mac = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{state}{\verb{<chr>} vector of state abbreviations}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search Physician Fee Schedule
}
\examples{
search_fee_schedule(hcpcs    = c("V5299", "70170"),
                    state    = "GA",
                    locality = "99",
                    mac      = "10212")

}
