% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_gpci}
\alias{search_gpci}
\title{Geographic Practice Cost Indices}
\usage{
search_gpci(mac = NULL, state = NULL, locality = NULL, ...)
}
\arguments{
\item{mac}{\verb{<chr>} 5-digit Medicare Administrative Contractor (MAC) ID}

\item{state}{\verb{<chr>} vector of state abbreviations}

\item{locality}{\verb{<chr>} vector of 2-digit Locality IDs}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Geographic Practice Cost Indices
}
\examples{
search_gpci(state = "GA", locality = "01", mac = "10212")

}
