% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_hcpcs}
\alias{search_hcpcs}
\title{HCPCS Level II Codes}
\usage{
search_hcpcs(hcpcs = NULL, columns = c("limit", "full"), ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{columns}{\verb{<chr>} set of columns returned, default is \code{limit}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
HCPCS Level II Codes
}
\examples{
search_hcpcs(hcpcs = c("A0021", "V5362", "J9264", "G8916"))

}
