% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_hcpcs}
\alias{search_hcpcs}
\title{Level II HCPCS Codes}
\usage{
search_hcpcs(hcpcs_code = NULL, ...)
}
\arguments{
\item{hcpcs_code}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
The Healthcare Common Procedure Coding System (HCPCS) is a collection of
codes that represent procedures, supplies, products and services which may be
provided to Medicare beneficiaries and to individuals enrolled in private
health insurance programs. The codes are divided into two levels, or groups,
as described below.
}
\details{
\itemize{
\item Level I Codes and descriptors copyrighted by the American Medical
Association's current procedural terminology, fourth edition (CPT-4). These
are 5 position numeric codes representing physician and non-physician
services.
\item Level II Includes codes and descriptors copyrighted by the American Dental
Association's current dental terminology, (CDT-2023). These are 5 position
alpha-numeric codes comprising the d series.
}

All level II codes and descriptors are approved and maintained jointly by the
alpha-numeric editorial panel (consisting of CMS, the Health Insurance
Association of America, and the Blue Cross and Blue Shield Association).

These are 5 position alpha-numeric codes representing primarily items and
non-physician services that are not represented in the level I codes.
}
\examples{
search_hcpcs(hcpcs_code = c("C9039", "J1835"))

}
\seealso{
Other HIPAA Standards: 
\code{\link{search_descriptions}()},
\code{\link{search_modifiers}()},
\code{\link{search_plas}()},
\code{\link{search_pos}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Standards}
