% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/describe.R
\name{search_level_two}
\alias{search_level_two}
\title{HCPCS Level II Codes}
\usage{
search_level_two()
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
HCPCS Level II Codes
}
\examples{
search_level_two()

}
