% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modifiers.R
\name{search_modifiers}
\alias{search_modifiers}
\title{Level I and II HCPCS Modifiers}
\usage{
search_modifiers(mod = NULL, ...)
}
\arguments{
\item{mod}{\emph{\if{html}{\out{<chr>}}} 2-digit HCPCS modifier}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
A modifier provides the means to report or indicate that a service or
procedure that has been performed has been altered by some specific
circumstance but not changed in its definition or code.
}
\details{
Modifiers also enable health care professionals to effectively respond to
payment policy requirements established by other entities.
}
\examples{
search_modifiers(mod = c("25", "59"))
}
