% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_payment}
\alias{search_payment}
\title{Search Payment Amount File}
\usage{
search_payment(hcpcs = NULL, mac = NULL, locality = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} 5-digit HCPCS code}

\item{mac}{\verb{<chr>} 5-digit MAC ID code}

\item{locality}{\verb{<chr>} 2-digit Locality ID}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search Payment Amount File
}
\examples{
search_payment(hcpcs    = c("39503", "43116", "33935", "11646"),
               locality = "01",
               mac      = "10212")
}
