% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standards.R
\name{search_pos}
\alias{search_pos}
\title{Place of Service (POS) Codes}
\usage{
search_pos(pos_code = NULL, pos_type = NULL, ...)
}
\arguments{
\item{pos_code}{\verb{<chr>} vector of 2-character Place of Service codes; default
is \code{NULL}}

\item{pos_type}{\verb{<chr>} Place of Service type, one of \code{Facility} or
\code{Non-Facility}; default is \code{NULL}}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Place of Service Codes are two-digit codes placed on health care professional
claims to indicate the setting in which a service was provided. The Centers
for Medicare & Medicaid Services (CMS) maintain POS codes used throughout the
health care industry.
}
\details{
This code set is required for use in the implementation guide adopted as the
national standard for electronic transmission of professional health care
claims under the provisions of the Health Insurance Portability and
Accountability Act of 1996 (HIPAA).

HIPAA directed the Secretary of HHS to adopt national standards for
electronic transactions. These standard transactions require all health plans
and providers to use standard code sets to populate data elements in each
transaction.

The Transaction and Code Set Rule adopted the ASC X12N-837 Health Care Claim:
Professional, volumes 1 and 2, version 4010, as the standard for electronic
submission of professional claims.

This standard names the POS code set currently maintained by CMS as the code
set to be used for describing sites of service in such claims. POS
information is often needed to determine the acceptability of direct billing
of Medicare, Medicaid and private insurance services provided by a given
provider.
}
\examples{
search_pos(pos_code = c("11", "21"))

search_pos(pos_type = "Facility")

}
\seealso{
Other HIPAA Code Standards: 
\code{\link{search_adjustments}()},
\code{\link{search_hcpcs}()},
\code{\link{search_modifiers}()},
\code{\link{search_rbcs}()}
}
\concept{HIPAA Code Standards}
