% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/databases.R
\name{search_rvu}
\alias{search_rvu}
\title{Search Relative Value File}
\usage{
search_rvu(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{< \emph{chr} > 5-digit HCPCS code}

\item{...}{Empty}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Search Relative Value File
}
\examples{
search_rvu(c("A0021", "V5362", "J9264", "G8916"))
}
