% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searches.R
\name{search_rvu}
\alias{search_rvu}
\title{Physician Fee Schedule Relative Value File}
\usage{
search_rvu(hcpcs = NULL, ...)
}
\arguments{
\item{hcpcs}{\verb{<chr>} vector of 5-character HCPCS codes}

\item{...}{These dots are for future extensions and must be empty.}
}
\value{
a \link[tibble:tibble-package]{tibble}
}
\description{
Physician Fee Schedule Relative Value File
}
\examples{
search_rvu(hcpcs = c("95907", "78140", "32820", "61575"))

}
