% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level2.R
\name{switch_action_code}
\alias{switch_action_code}
\title{HCPCS Level II Action Descriptions}
\usage{
switch_action_code(x)
}
\arguments{
\item{x}{vector of Action codes}
}
\value{
vector of Action descriptions
}
\description{
HCPCS Level II Action Descriptions
}
\examples{
dplyr::tibble(code = c(LETTERS), desc = switch_action_code(code))

}
