% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_asc_group}
\alias{switch_asc_group}
\title{HCPCS Level II ASC Group Descriptions}
\usage{
switch_asc_group(x)
}
\arguments{
\item{x}{vector of ASC Group indicators}
}
\value{
vector of ASC Group descriptions
}
\description{
HCPCS Level II ASC Group Descriptions
}
\examples{
dplyr::tibble(code = c("YY", "99"), desc = switch_asc_group(code))

}
