% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_assistant_surgery}
\alias{switch_assistant_surgery}
\title{PFS Assistant Surgery Descriptions}
\usage{
switch_assistant_surgery(x)
}
\arguments{
\item{x}{vector of Assistant Surgery indicators}
}
\value{
vector of Assistant Surgery descriptions
}
\description{
Fee schedule amount equals 16 percent of amount
otherwise applicable for surgical payment.
}
\details{
Modifiers:
\itemize{
\item 80: Assistance by Another Physician
\item 81: Minimal Assistance by a Another Physician
\item 82: Assistance by Another Physician when Qualified Resident Surgeon Unavailable
\item AS: Non-Physician Assistant at Surgery
}
}
\examples{
switch_assistant_surgery(as.character(0:9))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_diagnostic_imaging}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
