% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level2.R
\name{switch_betos}
\alias{switch_betos}
\title{HCPCS Level II BETOS Descriptions}
\usage{
switch_betos(x)
}
\arguments{
\item{x}{vector of BETOS codes}
}
\value{
vector of BETOS descriptions
}
\description{
HCPCS Level II BETOS Descriptions
}
\examples{
dplyr::tibble(code = c("D1A", "I2A", "M3", "P0"), desc = switch_betos(code))

}
