% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/case_switch.R
\name{switch_bilateral_surgery}
\alias{switch_bilateral_surgery}
\title{Bilateral Surgery Descriptions}
\usage{
switch_bilateral_surgery(x)
}
\arguments{
\item{x}{vector of Bilateral Surgery indicators}
}
\value{
vector of Bilateral Surgery descriptions
}
\description{
Modifier 50: Indicates services subject to payment adjustment
}
\examples{
switch_bilateral_surgery(c(0:3, "9"))

}
\keyword{internal}
