% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_co_surgeon}
\alias{switch_co_surgeon}
\title{Co-Surgeon Descriptions}
\usage{
switch_co_surgeon(x)
}
\arguments{
\item{x}{vector of Co-Surgeon indicators}
}
\value{
vector of Co-Surgeon descriptions
}
\description{
Modifier 62: Services for which two surgeons,
each in a different specialty, may be paid.
}
\examples{
switch_co_surgeon(c(0:2, 9, NA_character_))

}
