% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_level2.R
\name{switch_coverage}
\alias{switch_coverage}
\title{HCPCS Level II Coverage Descriptions}
\usage{
switch_coverage(x)
}
\arguments{
\item{x}{vector of Coverage indicators}
}
\value{
vector of Coverage descriptions
}
\description{
HCPCS Level II Coverage Descriptions
}
\examples{
dplyr::tibble(code = c("C", "D", "I", "M", "S"), desc = switch_coverage(code))

}
