% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_diagnostic_imaging}
\alias{switch_diagnostic_imaging}
\title{PFS Diagnostic Imaging Descriptions}
\usage{
switch_diagnostic_imaging(x)
}
\arguments{
\item{x}{vector of Diagnostic Imaging indicators}
}
\value{
vector of Diagnostic Imaging descriptions
}
\description{
Identifies the applicable Diagnostic Service family for
HCPCS codes with a Multiple Procedure indicator of 4.
}
\examples{
dplyr::tibble(
  code = fuimus::pad_number(c(1:11, 88, 99)),
  desc = switch_diagnostic_imaging(code))

}
\seealso{
Other PFS Helpers: 
\code{\link{switch_assistant_surgery}()},
\code{\link{switch_bilateral_surgery}()},
\code{\link{switch_co_surgeon}()},
\code{\link{switch_global_days}()},
\code{\link{switch_multiple_procedure}()},
\code{\link{switch_opps_indicator}()},
\code{\link{switch_pctc_indicator}()},
\code{\link{switch_status_description}()},
\code{\link{switch_status_name}()},
\code{\link{switch_supervision}()},
\code{\link{switch_team_surgery}()}
}
\concept{PFS Helpers}
