% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_global_days}
\alias{switch_global_days}
\title{PFS Global Days Descriptions}
\usage{
switch_global_days(x)
}
\arguments{
\item{x}{vector of Global Days indicators}
}
\value{
vector of Global Days descriptions
}
\description{
PFS Global Days Descriptions
}
\examples{
x <- c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ")
switch_global_days(x)

}
