% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_global_days}
\alias{switch_global_days}
\title{Global Days Descriptions}
\usage{
switch_global_days(x)
}
\arguments{
\item{x}{vector of Global Days indicators}
}
\value{
vector of Global Days descriptions
}
\description{
Global Days Descriptions
}
\examples{
switch_global_days(c("000", "010", "090", "MMM", "XXX", "YYY", "ZZZ"))

}
