% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_lab_certification}
\alias{switch_lab_certification}
\title{HCPCS Lab Certification Descriptions}
\usage{
switch_lab_certification(x)
}
\arguments{
\item{x}{vector of Lab Certification indicators}
}
\value{
vector of Lab Certification descriptions
}
\description{
HCPCS Lab Certification Descriptions
}
\examples{
dplyr::tibble(
  code = c("010", 100, 110, 115, 120, "110, 120, 130, 400"),
  desc = switch_lab_certification(code))

}
