% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_mue_mai}
\alias{switch_mue_mai}
\title{NCCI MUE Adjudication Indicator Descriptions}
\usage{
switch_mue_mai(x)
}
\arguments{
\item{x}{vector of MUE Adjudication Indicators}
}
\value{
vector of MUE Adjudication Indicator descriptions
}
\description{
NCCI MUE Adjudication Indicator Descriptions
}
\examples{
switch_mue_mai(as.character(1:3))

}
\seealso{
Other NCCI Helpers: 
\code{\link{switch_addon_edits}()},
\code{\link{switch_ptp_mod}()}
}
\concept{NCCI Helpers}
