% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switches.R
\name{switch_multiple_pricing}
\alias{switch_multiple_pricing}
\title{HCPCS Level II Multiple Pricing Descriptions}
\usage{
switch_multiple_pricing(x)
}
\arguments{
\item{x}{vector of Multiple Pricing indicators}
}
\value{
vector of Multiple Pricing descriptions
}
\description{
HCPCS Level II Multiple Pricing Descriptions
}
\examples{
dplyr::tibble(code = c(9, LETTERS[1:7]))

}
