% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_multiple_procedure}
\alias{switch_multiple_procedure}
\title{Multiple Procedure Descriptions}
\usage{
switch_multiple_procedure(x)
}
\arguments{
\item{x}{vector of Multiple Procedure indicators}
}
\value{
vector of Multiple Procedure descriptions
}
\description{
Modifier 51: Indicates applicable payment adjustment rule for multiple
procedures
}
\examples{
switch_multiple_procedure(as.character(0:9))

}
