% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_pctc_indicator}
\alias{switch_pctc_indicator}
\title{PC/TC Indicator Descriptions}
\usage{
switch_pctc_indicator(x)
}
\arguments{
\item{x}{vector of PC/TC indicators}
}
\value{
vector of PC/TC indicator descriptions
}
\description{
PC/TC Indicator Descriptions
}
\examples{
switch_pctc_indicator(as.character(0:9))

}
