% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_hcpcs.R
\name{switch_pricing}
\alias{switch_pricing}
\title{HCPCS Level II Pricing Descriptions}
\usage{
switch_pricing(x)
}
\arguments{
\item{x}{vector of Pricing indicators}
}
\value{
vector of Pricing descriptions
}
\description{
HCPCS Level II Pricing Descriptions
}
\examples{
dplyr::tibble(
  code = c("00", 11:13, 21:22, 31:46, 51:57, 99),
  desc = switch_pricing(code))

}
