% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_pfs.R
\name{switch_status_name}
\alias{switch_status_name}
\title{Status Code Names}
\usage{
switch_status_name(x)
}
\arguments{
\item{x}{vector of status codes}
}
\value{
vector of status code names
}
\description{
Status Code Names
}
\examples{
dplyr::tibble(
  code = LETTERS,
  name = switch_status_name(LETTERS),
  desc = switch_status_description(LETTERS)
)

}
