% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_hcpcs.R
\name{switch_supervision}
\alias{switch_supervision}
\title{HCPCS Level II Physician Supervision Descriptions}
\usage{
switch_supervision(x)
}
\arguments{
\item{x}{vector of Physician Supervision indicators}
}
\value{
vector of Physician Supervision descriptions
}
\description{
This field is for use in post payment review.
}
\examples{
dplyr::tibble(
  code = c(
    fuimus::pad_number(c(1:6, 9, 21:22, 66, 77)), "6A", "7A"),
  desc = switch_supervision(code))

}
