% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_fee.R
\name{switch_team_surgery}
\alias{switch_team_surgery}
\title{Team Surgery Descriptions}
\usage{
switch_team_surgery(x)
}
\arguments{
\item{x}{vector of Team Surgery indicators}
}
\value{
vector of Team Surgery descriptions
}
\description{
Modifier 66: Services for which Team Surgeons may be paid
}
\examples{
switch_team_surgery(c(0:2, "9"))

}
\keyword{internal}
