% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/switch_hcpcs.R
\name{switch_type_of_service}
\alias{switch_type_of_service}
\title{HCPCS Level II Type of Service Descriptions}
\usage{
switch_type_of_service(x)
}
\arguments{
\item{x}{vector of Type of Service codes}
}
\value{
vector of Type of Service descriptions
}
\description{
HCPCS Level II Type of Service Descriptions
}
\examples{
dplyr::tibble(code = c(0:9, LETTERS), desc = switch_type_of_service(code))

}
